% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_variants.R
\name{aac_to_sbs_ids}
\alias{aac_to_sbs_ids}
\title{Get sbs that cause an amino acid change}
\usage{
aac_to_sbs_ids(refcds_entry_name, aa_pos, aa_alt, bsg, refcds)
}
\arguments{
\item{aa_pos}{Integer position of substitution on the transcript.}

\item{aa_alt}{Identity of substitution, either a three-letter code ("Lys") or "STOP"}

\item{bsg}{A BSgenome object for the genome build associated with the RefCDS entry}

\item{refcds_entry}{A RefCDS entry for the relevant transcript}
}
\description{
An internal function to figure out the sbs that can cause a given amino acid substitution in a transcript
}
\keyword{internal}
