% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_handling.R
\name{add_clean_coverage_gr}
\alias{add_clean_coverage_gr}
\title{add_clean_coverage_gr}
\usage{
add_clean_coverage_gr(cesa, gr, covered_regions_name, coverage_type)
}
\arguments{
\item{cesa}{CESAnalysis to receive the gr}

\item{gr}{GRanges}

\item{covered_regions_name}{unique name for the covered regions}

\item{coverage_type}{"exome" or "targeted"}
}
\description{
Adds a validated GRanges object as a CESAnalysis's coverage set. Called by 
add_covered_regions() after various checks pass.
}
\details{
Special handling occurs if covered_regions_name is "exome+".
}
\keyword{internal}
