% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_handling.R
\name{add_covered_regions}
\alias{add_covered_regions}
\title{add_covered_regions}
\usage{
add_covered_regions(
  target_cesa = NULL,
  source_cesa = NULL,
  covered_regions = NULL,
  covered_regions_name = NULL,
  coverage_type = NULL,
  covered_regions_padding = 0
)
}
\arguments{
\item{target_cesa}{CESAnalysis that the covered regions will be added to}

\item{source_cesa}{Another CESAnalysis to copy all covered regions from}

\item{covered_regions}{A GRanges object or BED file path with genome build matching the target_cesa,
if not using source_cesa}

\item{covered_regions_name}{A name to identify the covered regions, if not using source_cesa}

\item{coverage_type}{exome or targeted, if not using source_cesa}

\item{covered_regions_padding}{optionally, add +/- this many bp to each interval in covered_regions}
}
\value{
CESAnalysis given in target_cesa, with the new covered regions added
}
\description{
add_covered_regions
}
