% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CESProgressions.R
\name{add_samples_to_CESProgressions}
\alias{add_samples_to_CESProgressions}
\title{add_samples_to_CESProgressions}
\usage{
add_samples_to_CESProgressions(progressions, samples, sample_progressions)
}
\arguments{
\item{progressions}{a CESProgressions object}

\item{samples}{a vector of unique patient identifiers}

\item{sample_progressions}{a vector of tumor progression stages corresonding to samples}
}
\value{
CESProgressions object with samples/progressions added
}
\description{
add_samples_to_CESProgressions
}
