% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_handling.R
\name{add_variants}
\alias{add_variants}
\title{Add variant annotations}
\usage{
add_variants(target_cesa = NULL, snv_id = NULL, source_cesa = NULL)
}
\arguments{
\item{target_cesa}{CESAnalysis to receive variant annotations}

\item{snv_id}{character vector of CES-style snv_ids to validate, annotate (including
finding and adding associated amino-acid-change mutations), and add to the target
CESAnalysis}

\item{source_cesa}{source CESAnalysis whose snv_ids will be used to create new
annotations in the target CESAnalysis}
}
\description{
Use this function to add variant annotations to your CESAnalysis, either using
a vector of CES-style snv_ids or a source CESAnalysis. Either way, the variants 
will be re-annotated using the target CESAnalysis's associated reference data. In other words,
when you use source_cesa, only the snv_ids are extracted, not any other information, in order
to avoid ugly situations involving conflicting reference data within a CESAnalysis.
}
