% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scna_helpers.R
\name{annotate_copy_change_consequence}
\alias{annotate_copy_change_consequence}
\title{Categorize copy changes relative to ploidy}
\usage{
annotate_copy_change_consequence(cna_calls)
}
\arguments{
\item{cna_calls}{calls}
}
\value{
A copy of the input table, with annotation columns added.
}
\description{
Categorize copy changes relative to ploidy
}
\keyword{internal}
