% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_snvs.R
\name{annotate_sbs}
\alias{annotate_sbs}
\title{Annotate SBS}
\usage{
annotate_sbs(sbs, refset)
}
\arguments{
\item{refset}{CES reference data set (e.g., from the ces.refset.hg38 data package).}

\item{variants}{MAF-like data.table of SBS.}
}
\description{
Annotates SBS with reference information, including protein changes.
}
\keyword{internal}
