% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_variants.R
\name{annotate_variants}
\alias{annotate_variants}
\title{Annotate variants}
\usage{
annotate_variants(cesa)
}
\arguments{
\item{cesa}{CESAnalysis object}
}
\description{
Annotates CESAnaysis MAF data with reference genome and gene data
}
