% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_signature_weights.R
\name{artifact_account}
\alias{artifact_account}
\title{Calculate relative rates of biological mutational processes}
\usage{
artifact_account(
  weights,
  signature_names,
  artifact_signatures = NULL,
  fail_if_zeroed = FALSE
)
}
\arguments{
\item{weights}{data.table of signature weights (can have extra columns)}

\item{signature_names}{names of signatures in weights (i.e., all column names)}

\item{artifact_signatures}{vector of artifact signature names (or NULL)}

\item{fail_if_zeroed}{T/F on whether to exit if a tumor would have all-zero weights.}
}
\description{
Sets artifact signature weights to zero and normalizes so that biologically-associated
weights sum to (1 - unattributed proportion) in each sample.
}
\keyword{internal}
