% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_rate_calc.R
\name{baseline_mutation_rates}
\alias{baseline_mutation_rates}
\title{baseline_mutation_rates}
\usage{
baseline_mutation_rates(cesa, aac_ids = NULL, snv_ids = NULL)
}
\description{
baseline_mutation_rates
}
