% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_RefCDS.R
\name{build_RefCDS}
\alias{build_RefCDS}
\title{buildref}
\usage{
build_RefCDS(
  gtf,
  genome,
  output_by = "gene",
  cds_ranges_lack_stop_codons = T,
  cores = 1,
  numcode = 1
)
}
\arguments{
\item{gtf}{Path of a Gencode-GTF-formatted text file, or an equivalently formatted data
table. See details for required columns (features). It's possible to build a suitable
table using data pulled from biomaRt, but it's probably easier to use a GTF.}

\item{genome}{genome assembly (e.g., "hg19"); an associated BSgenome object must be
available to load.}

\item{output_by}{"gene" (default) or "transcript", indicating whether the RefCDS output
should have one CDS annotation per gene or transcript.}

\item{cds_ranges_lack_stop_codons}{The CDS records in Gencode GTFs don't include the
stop codons in their genomic intervals. If your input table does include the stop 
codons within CDS records, set to FALSE.}

\item{cores}{how many cores to use for parallel computations}

\item{numcode}{(don't use) NCBI genetic code number; currently only code 1, the standard genetic code, is supported
  
Required columns are seqnames, start, end, strand, gene_name, gene_id, protein_id, and type. When
output_by = "transcript", there must also be a transcript_id column.
Only rows with type equal to "CDS" will be used. Strand should be "+" or "-".}
}
\description{
Function to build a dNdScv-style RefCDS object from a reference genome corresponding
transcript data. The dNdScv package This function generates an .rda file that needs to
be input into dndscv using the refdb argument. Note that when multiple CDS share the
same gene name (second column of cdsfile), the longest coding CDS will be chosen for
the gene. CDS with ambiguous bases (N) will not be considered.
}
\details{
Based on the buildref function in Inigo Martincorena's package dNdScv.
}
