% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scna_helpers.R
\name{calculate_ploidy}
\alias{calculate_ploidy}
\title{Determine even integer ploidy (2x, 4x, 8x)}
\usage{
calculate_ploidy(cna_calls)
}
\arguments{
\item{cna_calls}{calls}
}
\value{
A data.table of ploidy calls (2x, 4x, 8x).
}
\description{
Applies the approach used in Steele et al. to call sample ploidy.
We exclude chrX in samples not specifically annotated as female.
}
