% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_signature_weights.R
\name{calculate_trinuc_rates}
\alias{calculate_trinuc_rates}
\title{Calculate trinuc rates}
\usage{
calculate_trinuc_rates(weights, signatures, tumor_names)
}
\arguments{
\item{weights}{matrix of signature weights}

\item{signatures}{matrix of signatures}

\item{tumor_names}{names of tumors corresponding to rows of weights}
}
\value{
matrix of trinuc rates where each row corresponds to a tumor
}
\description{
Used interally to calculate trinuc rates from signature weights
}
\details{
If any relative rate is less than 1e-9, we add the lowest above-threshold rate to all
rates and renormalize rates so that they sum to 1.
}
\keyword{internal}
