% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_epistasis.R
\name{ces_epistasis}
\alias{ces_epistasis}
\title{Variant-level pairwise epistasis}
\usage{
ces_epistasis(
  cesa = NULL,
  variants = NULL,
  samples = character(),
  run_name = "auto",
  cores = 1,
  conf = 0.95,
  return_fit = FALSE
)
}
\arguments{
\item{cesa}{CESAnalysis}

\item{variants}{To test pairs of variants, supply a list where each element is a
2-length vector of CES-style variant IDs. Alternatively (and often more usefully),
supply a CompoundVariantSet (see \code{define_compound_variants}) to test all pairs
of compound variants in the set.}

\item{samples}{Which samples to include in inference. Defaults to all samples.
Can be a vector of Unique_Patient_Identifiers, or a data.table containing rows from
the CESAnalysis sample table.}

\item{run_name}{Optionally, a name to identify the current run.}

\item{cores}{number of cores for parallel processing of variant pairs}

\item{conf}{confidence interval size from 0 to 1 (.95 -> 95\%); NULL skips calculation,
reducing runtime.}

\item{return_fit}{Temporary option for whether to embed fitted models in a "fit" attribute associated
with the epistasis results table. This argument will be replaced soon with friendlier support 
for accessing fitted models.}
}
\value{
CESAnalysis with epistasis output in the form of a data.tabel of pairwise-epistatic
  selection intensities and variant frequencies for in tumors that have coverage at both variants
  in each pair
}
\description{
Calculate selection intensity under an assumption of pairwise epistasis between pairs of variants.
CompoundVariantSets are supported.
}
