% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_epistasis.R
\name{ces_epistasis_opm_args}
\alias{ces_epistasis_opm_args}
\title{Get recommended optimx optimization parameters for epistasis functions}
\usage{
ces_epistasis_opm_args()
}
\description{
Get recommended optimx optimization parameters for epistasis functions
}
\keyword{internal}
