% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_epistasis.R
\name{ces_gene_epistasis}
\alias{ces_gene_epistasis}
\title{Gene-level epistasis}
\usage{
ces_gene_epistasis(
  cesa = NULL,
  genes = character(),
  cores = 1,
  optimx_args = ces_epistasis_opm_args(),
  return_all_opm_output = FALSE
)
}
\arguments{
\item{cesa}{CESAnalysis object}

\item{genes}{vector of gene names; SIs will be calculated for all gene pairs}

\item{cores}{number of cores for parallel processing of gene pairs}

\item{optimx_args}{list of optimization parameters to feed into optimx::opm (see optimx docs); if you use this argument,
none of the CES default values will be used (opm defaults will be used for parameters you don't provide)}

\item{return_all_opm_output}{whether to include full parameter optimization report with output (default false)}
}
\value{
CESAnalysis object with selection results added for the chosen analysis
}
\description{
Calculate selection intensity under an assumption of pairwise gene-level epistasis.
}
