% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ces_gene_epistasis.R
\name{ces_gene_epistasis_opm_args}
\alias{ces_gene_epistasis_opm_args}
\title{Get list of arguments to feed into optimx::opm for parameter optimization during gene-level epistasis}
\usage{
ces_gene_epistasis_opm_args()
}
\description{
Get list of arguments to feed into optimx::opm for parameter optimization during gene-level epistasis
}
