% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_maf.R
\name{CESList}
\alias{CESList}
\title{Reads in an MAF file or data frame and creates a CESList object (the data structure used by cancereffectsizeR)}
\usage{
CESList(maf = NULL, sample_col = "Tumor_Sample_Barcode",
  chr_col = "Chromosome", start_col = "Start_Position",
  ref_col = "Reference_Allele", tumor_allele_col = "guess",
  consolidate_tcga_tumors_by_patient = TRUE)
}
\arguments{
\item{maf}{Path of tab-delimited text file in MAF format, or an MAF in data.frame format}

\item{sample_col}{column name with sample ID data}

\item{chr_col}{column name with chromosome data}

\item{start_col}{column name with start position ( currently only analyze SNVs, so no end position needed)}

\item{ref_col}{column name with reference allele data}

\item{tumor_allele_col}{column name with alternate allele data (defaults to "guess", which examines ref_col,
"Tumor_Seq_Allele1", and "Tumor_Seq_Allele2" and determines from there)}

\item{consolidate_tcga_tumors_by_patient}{logical: should apparent TCGA tumors be consolidated by patient?}
}
\description{
Reads in an MAF file or data frame and creates a CESList object (the data structure used by cancereffectsizeR)
}
