% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CESProgressions.R
\name{CESProgressions}
\alias{CESProgressions}
\title{CESProgressions}
\usage{
CESProgressions(samples, sample_progressions, ordered_progression_labels)
}
\arguments{
\item{ordered_progression_labels}{}
}
\description{
CESProgressions
}
