% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_data_handling.R
\name{check_for_genome_data}
\alias{check_for_genome_data}
\title{check_genome_data}
\usage{
check_for_genome_data(data_dir_or_cesa, datatype)
}
\description{
checks if the requested reference data exists and returns T/F
}
\keyword{internal}
