% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_data_handling.R
\name{check_for_ref_data}
\alias{check_for_ref_data}
\title{check_for_ref_data}
\usage{
check_for_ref_data(data_dir_or_cesa, datatype)
}
\arguments{
\item{data_dir_or_cesa}{CESAnalysis, or file path for reference data directory}
}
\description{
checks if the requested reference data exists and returns T/F
}
\keyword{internal}
