% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scna_helpers.R
\name{clean_rates}
\alias{clean_rates}
\title{Clean rates for use in current implementation of model}
\usage{
clean_rates(unclean_rates)
}
\description{
This function will eventually be removed (or at least, not exported)
}
