% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ces_variant.R
\name{clear_effect_output}
\alias{clear_effect_output}
\title{Clear variant effect output}
\usage{
clear_effect_output(cesa, run_names = names(cesa$selection))
}
\arguments{
\item{cesa}{CESAnalysis}

\item{run_names}{Which previous runs to remove; defaults to removing all.}
}
\description{
Remove output from previous ces_variant() runs from CESAnalysis
}
