% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ces_variant.R
\name{clear_epistasis_output}
\alias{clear_epistasis_output}
\title{Clear epistasis output}
\usage{
clear_epistasis_output(cesa, run_names = names(cesa$epistasis))
}
\arguments{
\item{cesa}{CESAnalysis.}

\item{run_names}{Which previous runs to remove; defaults to removing all.}
}
\description{
Remove previous epistatic effect estimations from CESAnalysis.
}
