% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_mutation_rates.R
\name{clear_gene_rates}
\alias{clear_gene_rates}
\title{Clear regional mutation rates}
\usage{
clear_gene_rates(cesa = NULL)
}
\arguments{
\item{cesa}{CESAnalysis}
}
\value{
The CESAnalysis with rates cleared
}
\description{
Remove all gene/coding region mutation rates, usually in order to re-run with different
parameters without having to create a new CESAnalysis.
}
