% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sample_data.R
\name{clear_sample_data}
\alias{clear_sample_data}
\title{Clear sample data}
\usage{
clear_sample_data(cesa, cols)
}
\arguments{
\item{cesa}{CESAnalysis}

\item{cols}{names of data columns to clear}
}
\description{
Remove data columns by name from CESAnalysis sample table. You can't clear
cancereffectsizeR-generated columns, such as coverage.
}
