% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trinuc_mutation_rates.R
\name{clear_trinuc_rates_and_signatures}
\alias{clear_trinuc_rates_and_signatures}
\title{Clear mutational signature attributions and related mutation rate information}
\usage{
clear_trinuc_rates_and_signatures(cesa)
}
\arguments{
\item{cesa}{CESAnalysis}
}
\description{
Removes all data calculated or supplied via trinuc_mutation_rates,
set_signature_weights, set_trinuc_rates, etc. This function can be used if you want to
re-run signature analysis with different sample groupings or parameters.
}
