% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_id_handling.R
\name{complete_aac_ids}
\alias{complete_aac_ids}
\title{Create full AAC ID}
\usage{
complete_aac_ids(partial_ids, refset)
}
\arguments{
\item{partial_ids}{AAC variant names, such as "KRAS_G12C" or "MIB2 G395C"}

\item{refset}{reference data set (environment object)}
}
\description{
For example, KRAS_G12C -> KRAS_G12C_ENSP00000256078 (ces.refset.hg19). In cases of
multiple protein IDs per gene, will return more IDs than input. Otherwise,
input/output will maintain order.
}
\details{
Output IDs are not fully validated. Use validate_aac_ids().
}
\keyword{internal}
