% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate_tcga_tumors_by_patient.R
\name{consolidate_tcga_tumors_by_patient}
\alias{consolidate_tcga_tumors_by_patient}
\title{Reads TCGA Tumor Sample Barcodes and consolidates barcodes into patient-specific
identifiers, such patients with multiple distinct tumors in the analysis share
share just one identifer. Sample barcodes that don't start with "TCGA" are unaltered.}
\usage{
consolidate_tcga_tumors_by_patient(
  tumor_sample_barcodes,
  non_TCGA_characters_to_keep = "all",
  sum_stats = T,
  figures = F
)
}
\arguments{
\item{tumor_sample_barcodes}{an MAF-style Tumor_Sample_Barcode vector, of which some or all are TCGA samples}

\item{non_TCGA_characters_to_keep}{numeric number of characters to keep
in "Unique_Patient_Identifier" column for sample names that
do not start with "TCGA". If non-numeric object then defaults to
all characters}

\item{sum_stats}{Boolean, T prints summary() of number of mutations per tumor}

\item{figures}{Boolean, T saves histogram of sum.stats and saves as "Figures/mutations_per_tumor.pdf"}
}
\description{
Reads TCGA Tumor Sample Barcodes and consolidates barcodes into patient-specific
identifiers, such patients with multiple distinct tumors in the analysis share
share just one identifer. Sample barcodes that don't start with "TCGA" are unaltered.
}
