% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_signature_weights.R
\name{convert_signature_weights_for_mp}
\alias{convert_signature_weights_for_mp}
\title{Get MutationalPatterns contributions matrix}
\usage{
convert_signature_weights_for_mp(signature_weight_table)
}
\arguments{
\item{signature_weight_table}{As created by trinuc_mutation_rates(); typically accessed
via (CESAnalysis)$mutational_signatures.}
}
\description{
Reformat a signature weights table from mutational signature analysis into the
contributions matrix required for MutationalPatterns functions, including
visualizations.
}
