% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CESAnalysis.R
\name{copy_cesa}
\alias{copy_cesa}
\title{Creat an independent copy of a CESAnalysis}
\usage{
copy_cesa(cesa)
}
\arguments{
\item{cesa}{CESAnalysis}
}
\description{
Used internally to "copy" CESAnalysis objects while keeping memory use to a minimum.
}
\details{
The trick is to use data.table's copy function on all data.tables (and lists of
data.tables) within CESAnalysis slots. (If you just call copy on the whole object, the
data tables won't be handled in a memory-efficient way. And if you call copy on
non-data.tables, it's actuall less-efficient since it forces immediate copy instead of
the usual copy-on-modify.)
}
\keyword{internal}
