% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_signature_effects.R
\name{cosmic_sbs_etiologies}
\alias{cosmic_sbs_etiologies}
\title{COSMIC signatures with known etiologies}
\usage{
cosmic_sbs_etiologies()
}
\description{
COSMIC signatures with known etiologies
}
