% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_signature_effects.R
\name{cosmic_signature_info}
\alias{cosmic_signature_info}
\title{Get COSMIC signature descriptions}
\usage{
cosmic_signature_info()
}
\description{
Returns a table describing COSMIC signatures. All signatures from v3.0 to the latest release
(v3.4) are included, with information derived from the most recent information on the COSMIC
website. (Exception: A reported association between SBS16 and alcohol consumption, noted here, is
not mentioned on the COSMIC website.)
}
