% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_epistasis_pvalue_calc.R
\name{default_epistasis_pvalue_calc}
\alias{default_epistasis_pvalue_calc}
\title{Calculate epistasis p-values}
\usage{
default_epistasis_pvalue_calc(fit)
}
\arguments{
\item{fit}{fitted epistatic model}
}
\description{
Calculates p-values and other interesting information.
}
\keyword{internal}
