% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_handling.R
\name{define_compound_variants}
\alias{define_compound_variants}
\title{Clump variants together}
\usage{
define_compound_variants(cesa, snv_id)
}
\arguments{
\item{cesa}{CESAnalysis, just used to validate the snv_id against reference data.}

\item{snv_id}{Vector of snv_ids to include in one compound variant, or a list of
vectors, each of which defines a separate compound variant. If the vector or list is
named, names will be kept. Otherwise, compound variants will be named sequentially}
}
\description{
Details coming
}
