% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_mnv.R
\name{detect_mnv}
\alias{detect_mnv}
\title{Find likely MNVs in an MAF table}
\usage{
detect_mnv(maf)
}
\arguments{
\item{maf}{a valid MAF-style data.table}
}
\value{
a table with MAF records
}
\description{
Same-sample variants with 2bp of other variants are likely MNVs.
}
\keyword{internal}
