% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_mutation_rates.R
\name{dndscv_postprocess}
\alias{dndscv_postprocess}
\title{Internal function to calculate gene-level mutation rates from dNdScv output}
\usage{
dndscv_postprocess(cesa, dndscv_raw_output, save_all_dndscv_output = FALSE)
}
\description{
Internal function to calculate gene-level mutation rates from dNdScv output
}
