% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_level_mutation_rates.R
\name{dndscv_preprocess}
\alias{dndscv_preprocess}
\title{Internal function to prepare for running dNdScv}
\usage{
dndscv_preprocess(cesa, covariate_file = NULL)
}
\description{
Internal function to prepare for running dNdScv
}
