% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNP_TNP_remover.R
\name{DNP_TNP_remover}
\alias{DNP_TNP_remover}
\title{Remove likely multi-nucleotide mutation events
 
 Function that removes DNP that are actually labeled as single nucleotide variants.}
\usage{
DNP_TNP_remover(MAF, delete_recur = F)
}
\arguments{
\item{MAF}{MAF file, with relevant fields consisten with the GDC MAF specification (\url{https://docs.gdc.cancer.gov/Data/File_Formats/MAF_Format/}).}

\item{delete_recur}{T/F statement about whether to remove recurrent (02) tumors.}
}
\description{
Remove likely multi-nucleotide mutation events
 
 Function that removes DNP that are actually labeled as single nucleotide variants.
}
\keyword{internal}
