% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_counts.R
\name{.variant_counts}
\alias{.variant_counts}
\title{Internal variant prevalence and coverage calculation}
\usage{
.variant_counts(
  cesa,
  samples,
  snv_from_aac,
  noncoding_snv_id,
  by_cols = character()
)
}
\arguments{
\item{cesa}{CESAnalysis}

\item{samples}{validated samples table}

\item{snv_from_aac}{data.table with columns aac_id, snv_id (validated and with annotations in CESAnalysis)}

\item{noncoding_snv_id}{vector of snv_ids to treat as noncoding variants}

\item{by_cols}{validated column names from sample table that are suitable to use for counting by.}
}
\description{
Called by variant_counts() (and select_variants()) with validated inputs.
}
\keyword{internal}
