% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_epistasis.R
\name{epistasis_plot_schematic}
\alias{epistasis_plot_schematic}
\title{Get epistatic effect schematic}
\usage{
epistasis_plot_schematic(
  title = "Types of effects",
  schematic_label_size = 3,
  with_border = TRUE
)
}
\arguments{
\item{title}{Schematic title text.}

\item{schematic_label_size}{Text size of labels in the schematic (title gets size + 1).}

\item{with_border}{TRUE/FALSE on the appearance of a thin visible border around the schematic.}
}
\value{
The schematic as a ggplot.
}
\description{
Get a copy of the explanatory schematic that appears in epistatic effect plots (see
\code{plot_epistasis()}). May be useful for putting the schematic in custom locations when
assembling complex figures.
}
