% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CESAnalysis.R
\name{epistasis_results}
\alias{epistasis_results}
\title{View output from epistasis functions}
\usage{
epistasis_results(cesa = NULL)
}
\arguments{
\item{cesa}{CESAnalysis object}
}
\description{
returns a list of data tables with results from epistasis functions
}
