% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CESAnalysis.R
\name{excluded_maf_records}
\alias{excluded_maf_records}
\title{excluded_maf_records}
\usage{
excluded_maf_records(cesa = NULL)
}
\arguments{
\item{cesa}{CESAnalysis object}
}
\description{
returns a data.table containing MAF records that were excluded from the given CESAnalysis
}
