% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_mutation_rates.R
\name{fill_missing_genes}
\alias{fill_missing_genes}
\title{Fill missing genes}
\usage{
fill_missing_genes(gr_genes, gene_rates)
}
\arguments{
\item{gr_genes}{reference data set style gr_genes}

\item{gene_rates}{two-column data.table (gene, rate)}
}
\description{
For each gene in gr_genes missing from gene rate table, assign the 
rate of the nearest non-missing gene. Gene rate table is assumed
two-column (gene and rate).
}
\keyword{internal}
