% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CESAnalysis.R
\name{gene_epistasis_results}
\alias{gene_epistasis_results}
\title{View results from gene-level epistasis analysis}
\usage{
gene_epistasis_results(cesa = NULL)
}
\arguments{
\item{cesa}{CESAnalysis object}
}
\description{
returns a data table of pairwise gene epistasis effect sizes generated with ces_gene_epistasis
}
