% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_level_mutation_rates.R
\name{gene_level_mutation_rates}
\alias{gene_level_mutation_rates}
\title{Gene-level mutation rates}
\usage{
gene_level_mutation_rates(
  cesa,
  covariate_file = NULL,
  save_all_dndscv_output = FALSE
)
}
\arguments{
\item{cesa}{CESAnalysis object}

\item{covariate_file}{Either NULL and usefs dNdScv default covariates, or one of these:  "bladder_pca"  "breast_pca"
"cesc_pca" "colon_pca" "esca_pca" "gbm_pca" "hnsc_pca" "kidney_pca" "lihc_pca" "lung_pca" "ov_pca" "pancreas_pca" "prostate_pca" "rectum_pca" "skin_pca"  "stomach_pca"  "thca_pca" "ucec_pca"}

\item{save_all_dndscv_output}{default false; when true, saves all dndscv output, not just what's needed by CES}
}
\value{
CESAnalysis object with gene-level mutation rates calculated
}
\description{
This function calculates gene-level neutral mutation rates based on counts
of nonsynonymous and synonymous mutations per gene under the dNdScv package's model, 
as described in Martincorena et al. (https://doi.org/10.1016/j.cell.2017.09.042).
}
