% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_mutation_rates.R
\name{gene_mutation_rates}
\alias{gene_mutation_rates}
\title{Use dNdScv with tissue-specific covariates to calculate gene-level mutation rates}
\usage{
gene_mutation_rates(
  cesa,
  covariates = NULL,
  sample_group = NULL,
  save_all_dndscv_output = FALSE
)
}
\arguments{
\item{cesa}{CESAnalysis object}

\item{covariates}{Name of tissue-specific covariates to use (run list_ces_covariates()
to see choices). For hg19 data only, you can also use "default" for dNdScv's
non-tissue-specific covariates. If no appropriate covariates data are available, set
NULL to run without. Finally, you can also supply custom covariates data in the form
of a prcomp object (see documentation for details).}

\item{sample_group}{Which sample groups to include in the gene rate calculation;
defaults to all groups. (To calculate different rates for different groups, you'll 
run this function multiple times, changing this argument each time.)}

\item{save_all_dndscv_output}{default false; when true, saves all dndscv output, not
just what's needed by cancereffectsizeR (will make object size very large)}
}
\value{
CESAnalysis object with gene-level mutation rates calculated
}
\description{
This function calculates gene-level neutral mutation rates based on counts
of nonsynonymous and synonymous mutations per gene under the dNdScv package's model, 
as described in \href{https://doi.org/10.1016/j.cell.2017.09.042}{Martincorena et al.}
}
