% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_data_handling.R
\name{get_cesa_bsg}
\alias{get_cesa_bsg}
\title{get_cesa_bsg}
\usage{
get_cesa_bsg(cesa)
}
\arguments{
\item{cesa}{CESAnalysis}
}
\description{
Loads the right BSgenome for a CESAnalysis
}
\keyword{internal}
