% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scna_helpers.R
\name{get_chr_change_counts}
\alias{get_chr_change_counts}
\title{Tabulate copy changes by genomic window.}
\usage{
get_chr_change_counts(calls, chr, chr_ranges, account = TRUE)
}
\arguments{
\item{chr}{Chromosome name as 1-length character}

\item{chr_ranges}{From get_genomic_windows()}

\item{account}{When TRUE (default), When account = TRUE, arm-level and chr-level changes are 
accounted for (i.e., ignored) when counting changes.}
}
\description{
Across a chromosome, counts the number of samples with gains, decreases, LOH, non-LOH decrease,
etc. All changes are with respect to expected ploidy.
}
