% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_mutation_rates.R
\name{get_dndscv_model_fit}
\alias{get_dndscv_model_fit}
\title{This little function called by gene_mutation_rates() is separated for testing purposes.}
\usage{
get_dndscv_model_fit(dndscv_output)
}
\description{
This little function called by gene_mutation_rates() is separated for testing purposes.
}
\keyword{internal}
