% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CESAnalysis.R
\name{get_gene_rates}
\alias{get_gene_rates}
\title{Get table of neutral gene mutation rates}
\usage{
get_gene_rates(cesa = NULL)
}
\arguments{
\item{cesa}{CESAnalysis object}
}
\description{
Get table of neutral gene mutation rates
}
