% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_data_handling.R
\name{get_genome_data}
\alias{get_genome_data}
\title{get_genome_data}
\usage{
get_genome_data(data_dir_or_cesa, datatype)
}
\description{
reads in the requested reference data for the genome build associated with the CESAnalysis
}
\keyword{internal}
