% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_genome.R
\name{get_genome_dirs}
\alias{get_genome_dirs}
\title{get_genome_dirs
returns a character vector with the paths of genome data available to use with CES (or just genome names)
does not verify that all required files are present in the genome directories}
\usage{
get_genome_dirs(full_paths = TRUE)
}
\arguments{
\item{full_paths}{default TRUE; returns full paths to genome data instead of just genome names}
}
\description{
get_genome_dirs
returns a character vector with the paths of genome data available to use with CES (or just genome names)
does not verify that all required files are present in the genome directories
}
