% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genomic_windows.R
\name{get_genomic_windows}
\alias{get_genomic_windows}
\title{Split genome into same-width ranges.}
\usage{
get_genomic_windows(
  refset,
  window_size = 1e+05,
  ranges = NULL,
  exclude_ENCODE_blacklist = TRUE,
  exclude_non_PAR = TRUE
)
}
\arguments{
\item{window_size}{Integer size of windows. Will be a little slow if windows are very small.}

\item{ranges}{If supplied, will only split up the input GRanges, as opposed to the full primary assembly reference ranges.}

\item{exclude_ENCODE_blacklist}{TRUE/FALSE on filtering out windows that overlap large (> .5 * window_size)
regions flagged in ENCODE blacklist v2 (default TRUE).}

\item{exclude_non_PAR}{TRUE/FALSE on removing windows that overlap non-pseudoautosomal regions of chrX/chrY (default TRUE).}
}
\description{
## TO-DO: change to excluding windows that are >50% blacklisted, instead of the weird criteria here.
}
