% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_variants.R
\name{get_gr_from_table}
\alias{get_gr_from_table}
\title{Get GRanges from chr/start/end table}
\usage{
get_gr_from_table(variant_table)
}
\arguments{
\item{variant_table}{data.table}
}
\description{
Mainly built for select_variants() output, and uses the
center_nt_pos on AACs (rather than all from start-end). Assumes
MAF-like coordinates (1-based, closed).
}
\keyword{internal}
