% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CESAnalysis.R
\name{get_mutations}
\alias{get_mutations}
\title{Get lists of mutations and annotations}
\usage{
get_mutations(cesa = NULL)
}
\arguments{
\item{cesa}{CESAnalysis object}
}
\description{
Get lists of mutations and annotations
}
