% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_variants.R
\name{get_nearest_refset_entries}
\alias{get_nearest_refset_entries}
\title{Annotates MAF-like table with nearest gene and CDS entries}
\usage{
get_nearest_refset_entries(variants, refset)
}
\arguments{
\item{variants}{MAF-like data.table}

\item{refset}{CES reference data environment}
}
\description{
Annotates MAF-like table with nearest gene and CDS entries
}
\keyword{internal}
