% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_PathScore_input.R
\name{get_PathScore_coding_regions}
\alias{get_PathScore_coding_regions}
\title{Get PathScore coding regions}
\usage{
get_PathScore_coding_regions(genome = "hg38")
}
\arguments{
\item{genome}{Genome build: Either "hg39" (default) or "hg19".}
}
\description{
Returns GRanges that represent the coding sequence (CDS) definitions used by PathScore. The hg19
version was created by running liftOver on the hg38 intervals.
}
