% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_data_handling.R
\name{get_ref_data}
\alias{get_ref_data}
\title{get_ref_data}
\usage{
get_ref_data(data_dir_or_cesa, datatype)
}
\description{
reads in the requested reference data for the ref set associated with the CESAnalysis
}
\keyword{internal}
