% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_data_handling.R
\name{get_refset_dirs}
\alias{get_refset_dirs}
\title{get_refset_dirs}
\usage{
get_refset_dirs()
}
\description{
returns a character vector mapping ref set names to their data directories
}
\keyword{internal}
