% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CESAnalysis.R
\name{get_sample_info}
\alias{get_sample_info}
\title{View sample metadata}
\usage{
get_sample_info(cesa = NULL)
}
\arguments{
\item{cesa}{CESAnalysis object}
}
\description{
returns a data.table with info on all samples in the CESAnalysis
}
