% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CESAnalysis.R
\name{get_signature_weights}
\alias{get_signature_weights}
\title{Get table of signature weights by tumor}
\usage{
get_signature_weights(cesa = NULL)
}
\arguments{
\item{cesa}{CESAnalysis object}
}
\description{
Get table of signature weights by tumor
}
