% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tumors_with_coverage.R
\name{get_tumors_with_coverage}
\alias{get_tumors_with_coverage}
\title{get_tumors_with_coverage}
\usage{
get_tumors_with_coverage(coverage, locus)
}
\arguments{
\item{coverage}{list as formatted in the coverage slot of a CESAnalysis}

\item{locus}{a single locus (chr:start-end) in a GRanges object}
}
\value{
a character vector of tumor identifiers
}
\description{
get_tumors_with_coverage
}
