% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hg_converter.R
\name{hg_converter}
\alias{hg_converter}
\title{hg_converter}
\usage{
hg_converter(chain, maf_to_convert, chrom_col_name = "Chromosome",
  new_build_name = "Converted_from_GRCh38_to_hg19")
}
\arguments{
\item{chain}{A directory on your system with the "*over.chain" file to be used with liftOver. For instance, the chain to convert from hg38 to hg19 can to be downloaded from UCSC \url{http://hgdownload.cse.ucsc.edu/gbdb/hg38/liftOver/} (hg38ToHg19.over.chain.gz).}

\item{maf_to_convert}{MAF file, with relevant fields consisten with the GDC MAF specification (\url{https://docs.gdc.cancer.gov/Data/File_Formats/MAF_Format/}), used in the conversion.}

\item{chrom_col_name}{Name of the column header containing chromosome information.}

\item{new_build_name}{This will be the string in the "NCBI_Build" column header of the new MAF file.}
}
\value{
The returned output will be a MAF file with all the same data fields as the original but with "Start_Position" and "End_Position" converted to the build specified by the 'chain' input.
}
\description{
Convert MAF files from one genomic coordinate system to another.
Developed for use with MAF files found on the
NCI Genome Data Commons. Adapted from the liftOver vignettes page at \url{https://master.bioconductor.org/packages/release/workflows/vignettes/liftOver/inst/doc/liftov.html}
}
