% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_read_maf.R
\name{identify_maf_variants}
\alias{identify_maf_variants}
\title{Annotate MAF data with variant types and variant IDs}
\usage{
identify_maf_variants(maf)
}
\arguments{
\item{maf}{a validated data.table with MAF-like columns}
}
\value{
input table with variant_type and variant_id columns
}
\description{
Annotate MAF data with variant types and variant IDs
}
\keyword{internal}
