% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lift_bed.R
\name{lift_bed}
\alias{lift_bed}
\title{Convert BED intervals between genome builds}
\usage{
lift_bed(bed, chain, outfile = NULL)
}
\arguments{
\item{bed}{Pathname of a BED file, or a GRanges (typically loaded from a BED file with \code{rtracklayer::import.bed()}).}

\item{chain}{A UCSC-style chain file, or a Chain object (such as from \code{rtracklayer::import.chain()}).}

\item{outfile}{If not NULL, the returned GRanges will be saved to the specified path using \code{rtracklayer::export.chain()}.}
}
\value{
GRanges representing lifted intervals from input \code{bed}.
}
\description{
Use this utility to convert BED intervals between genome coordinate systems using liftOver. Only
the chr/start/end fields of the input BED are used (strand is ignored). The output GRanges
will have no associated seqinfo.
}
\details{
A warning is given if the lifted intervals are less than 95\% of the size of the original
intervals. When the BED input represents sequencing target intervals, most of the input
intervals will usually lift successfully.
}
