% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_data_handling.R
\name{list_ces_covariates}
\alias{list_ces_covariates}
\title{list_ces_covariates}
\usage{
list_ces_covariates()
}
\description{
Prints names of available built-in covariate data sets for all loaded CES genomes
}
