% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_data_handling.R
\name{list_ces_genomes}
\alias{list_ces_genomes}
\title{list_ces_genomes}
\usage{
list_ces_genomes()
}
\description{
Prints names of reference data collections that are ready for use with cancereffectsizeR
}
