% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_data_handling.R
\name{list_ces_signature_sets}
\alias{list_ces_signature_sets}
\title{list_ces_signature_sets}
\usage{
list_ces_signature_sets()
}
\description{
Prints names of available mutational signature sets. Just to be clear, we're calling
them ces_signature_sets because they're ready to use with cancereffectsizeR. We didn't
derive any of these signature sets.
}
