% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_genome.R
\name{list_genomes}
\alias{list_genomes}
\title{prints names of  all genome builds that are ready for use with cancereffectsizeR
list_genomes}
\usage{
list_genomes()
}
\description{
prints names of  all genome builds that are ready for use with cancereffectsizeR
list_genomes
}
