% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CESAnalysis.R
\name{load_cesa}
\alias{load_cesa}
\title{Load a previously saved CESAnalysis object}
\usage{
load_cesa(file)
}
\arguments{
\item{file}{filename/path of CESAnalysis that has been saved with saveRDS, expected to end in .rds}
}
\description{
Load a previously saved CESAnalysis object
}
