% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sample_data.R
\name{load_sample_data}
\alias{load_sample_data}
\title{Add sample data}
\usage{
load_sample_data(cesa, sample_data)
}
\arguments{
\item{cesa}{CESAnalysis}

\item{sample_data}{data.table or data.frame with a Unique_Patient_Identifier column to
match the CESAnalysis samples table, with one row per sample. (It's okay if some
samples aren't present in the table.) For simplicity, columns besides
Unique_Patient_Identifier must not already be present in the CESAnalysis. (Instead,
build a single sample_data table and load it with one call to this function.)}
}
\description{
Insert sample-level data into a CESAnalysis samples table.
}
