% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CESAnalysis.R
\name{maf_records}
\alias{maf_records}
\title{View data loaded into CESAnalysis}
\usage{
maf_records(cesa = NULL)
}
\arguments{
\item{cesa}{CESAnalysis object}
}
\description{
returns a data.table containing MAF records used in the given CESAnalysis
}
