% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{MAFdf-class}
\alias{MAFdf-class}
\title{A class that extends data.frame to enforce consistent handling of MAF data}
\description{
A class that extends data.frame to enforce consistent handling of MAF data
}
