% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scna_helpers.R
\name{make_biscut_chr_info}
\alias{make_biscut_chr_info}
\title{Make a BISCUT-style chromosome info data.frame}
\usage{
make_biscut_chr_info(
  arm_coordinates,
  acrocentromeric_chr = c("13", "14", "15", "21", "22")
)
}
\arguments{
\item{arm_coordinates}{data.table that defines the p/q arms of each chromosome.}

\item{acrocentromeric_chr}{character vector with the names of chromosomes that are acrocentromeric}
}
\value{
A data.frame of the sort favored by BISCUT::validate_chr_coordinates().
}
\description{
Make a BISCUT-style chromosome info data.frame
}
