% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scna_helpers.R
\name{make_biscut_input}
\alias{make_biscut_input}
\title{Convert SCNA calls into BISCUT's format.}
\usage{
make_biscut_input(calls)
}
\description{
As stated in the BISCUT REAMDE: "If you have your data in absolute allelic copy number format, to
 obtain Segment_Mean for each segment, simply add the allelic copy numbers and then divide that by
 the ploidy of the sample and take log2."
 
 @param calls SCNA segment calls, as formatted by prep_ASCAT3_segments()
 @return Data.table suitable for use with BISCUT::calculate_breakpoints(). Note that chrY is
   excluded, and chrX is relabeled as "23". The Num_Probes field is set to 10 for non-probe data,
   as instructed in BISCUT documentation.
 @export
}
