% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merging_NCI_and_local_MAF_files.R
\name{merging_TCGA_and_local_MAFdata_function}
\alias{merging_TCGA_and_local_MAFdata_function}
\title{Merging TCGA and local MAF data}
\usage{
merging_TCGA_and_local_MAFdata_function(
  NCI_data,
  Local_data,
  check_for_same_tumor = T
)
}
\arguments{
\item{NCI_data}{MAF downloaded from the NCI GDC}

\item{Local_data}{MAF created locally}

\item{check_for_same_tumor}{Boolean, if T it checks how many substitutions
are common among each "Local_data" tumor and all "NCI_data" tumors
in an attempt to inform the user of potential duplicated tumors (e.g. 
tumor obtained from a lab was already incorporated within data 
stored on NCI GDC and it is the same tumor with a different ID). 
If a "Local" tumor shares >10% of the same SNV, a file 
"potential_duplicates.RData" is saves to the working directory with 
potential duplicates data.}
}
\description{
This function merges data frames that are both in "MAF" format
(i.e. \url{https://docs.gdc.cancer.gov/Data/File_Formats/MAF_Format/}), and
is meant to merge a MAF from the NCI (with all correct headers)
with a MAF created locally, and attempts to merge along 
common differences within column names (check that your 
column names are in MAF format!)
}
