% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objective_epistasis.R
\name{ml_objective_epistasis}
\alias{ml_objective_epistasis}
\title{ml_objective for epistasis calculation}
\usage{
ml_objective_epistasis(
  gamma,
  MAF_input1,
  MAF_input2,
  all_tumors,
  variant1,
  variant2,
  specific_mut_rates1,
  specific_mut_rates2
)
}
\arguments{
\item{gamma}{A selection intensity at which to calculate the likelihood}

\item{MAF_input1}{A data frame that includes columns "Unique_patient_identifier", "Gene_name", and "unique_variant_ID_AA"}

\item{MAF_input2}{A data frame that includes columns "Unique_patient_identifier", "Gene_name", and "unique_variant_ID_AA"}

\item{all_tumors}{A list of all the tumors we are calculating the likelihood across}

\item{variant1}{one variant in epistatic analysis}

\item{variant2}{the other}

\item{specific_mut_rates1}{site and tumor specific mutation rates (matrix: rows = tumors, columns = variants; produced by mutation_rate_calc)}

\item{specific_mut_rates2}{site and tumor specific mutation rates (matrix: rows = tumors, columns = variants; produced by mutation_rate_calc)}
}
\value{
A log likelihood value
}
\description{
Objective function that we will be optimizing in order to find the site specific selection intensity that maximizes the likelihood of each tumor having a mutation or not, where the mutation rates are site and tumor specific.
}
