% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objective_epistasis_full_gene.R
\name{ml_objective_epistasis_full_gene}
\alias{ml_objective_epistasis_full_gene}
\title{ml_objective for epistasis calculation at the gene level}
\usage{
ml_objective_epistasis_full_gene(gamma, MAF_input1, MAF_input2, all_tumors,
  gene1, gene2, specific_mut_rates1, specific_mut_rates2)
}
\arguments{
\item{gamma}{A selection intensity at which to calculate the likelihood}

\item{all_tumors}{A list of all the tumors we are calculating the likelihood across}

\item{MAF_input}{A data frame that includes columns "Unique_patient_identifier", "Gene_name", and "unique_variant_ID_AA"}

\item{gene}{The gene we want to look at}

\item{variant}{The variant we want to look at}

\item{specific_mut_rates}{A matrix of site and tumor specific mutation rates where the rows correspond to tumors and the columns to variants (produced by mutation_rate_calc)}
}
\value{
A log likelihood value
}
\description{
Objective function that we will be optimizing in order to find the site specific selection intensity that maximizes the likelihood of each tumor having a mutation or not, where the mutation rates are site and tumor specific.
}
