% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objective_pairwise_epistasis.R
\name{ml_objective_pairwise_epistasis}
\alias{ml_objective_pairwise_epistasis}
\title{ml_objective for pairwise epistasis between variants or genes}
\usage{
ml_objective_pairwise_epistasis(
  with_just_1,
  with_just_2,
  with_both,
  with_neither
)
}
\arguments{
\item{with_just_1}{two-item list of baseline rates in v1/v2 for tumors with mutation in just the first variant(s)}

\item{with_just_2}{two-item list of baseline rates in v1/v2 for tumors with mutation in just the second variant(s)}

\item{with_both}{two-item list of baseline rates in v1/v2 for tumors with mutation in both}

\item{with_neither}{two-item list of baseline rates in v1/v2 for tumors with mutation n neither}

\item{par}{A selection intensity at which to calculate the likelihood}
}
\value{
A log likelihood value
}
\description{
Objective function that we will be optimizing (minimizing the negative of the log likelihood, so maximizing the log likelihood) 
in order to find the site specific selection intensity that maximizes the likelihood of each tumor having a mutation or not, 
where the mutation rates are site and tumor specific.
}
\keyword{internal}
