% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_finder.R
\name{mutation_finder}
\alias{mutation_finder}
\title{mutation_finder}
\usage{
mutation_finder(RefCDS_instance, MAF_input_row)
}
\arguments{
\item{RefCDS_instance}{Slice of the RefCDS list for this gene}

\item{MAF_input_row}{Row from the MAF being analyzed with the mutation in question.}
}
\value{

}
\description{
Finds original trinucleotides of variants near splice sites.
}
