% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_rate_calc.R
\name{mutation_rate_calc}
\alias{mutation_rate_calc}
\title{mutation_rate_calc}
\usage{
mutation_rate_calc(MAF, gene, gene_mut_rate, tumor_trinucs)
}
\arguments{
\item{tumor_trinucs}{}
}
\description{
mutation_rate_calc
}
