% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_gamma.R
\name{optimize_gamma}
\alias{optimize_gamma}
\title{optimize_gamma}
\usage{
optimize_gamma(MAF_input, all_tumors, gene, variant, specific_mut_rates)
}
\arguments{
\item{specific_mut_rates}{}
}
\description{
optimize_gamma
}
