% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_gamma_epistasis.R
\name{optimize_gamma_epistasis}
\alias{optimize_gamma_epistasis}
\title{Calculate selection intensity and epistasis}
\usage{
optimize_gamma_epistasis(
  MAF_input1,
  MAF_input2,
  all_tumors,
  gene1,
  gene2,
  variant1,
  variant2,
  specific_mut_rates1,
  specific_mut_rates2
)
}
\arguments{
\item{all_tumors}{A list of all the tumors we are calculating the likelihood across}

\item{MAF_input}{A data frame that includes columns "Unique_patient_identifier", "Gene_name", and "unique_variant_ID_AA"}

\item{gene}{The gene we want to look at}

\item{variant}{The variant we want to look at}

\item{specific_mut_rates}{A matrix of site and tumor specific mutation rates where the rows correspond to tumors and the columns to variants (produced by mutation_rate_calc)}
}
\value{
The optimal selection intensity for the gene and variant that maximizes the likelihood of the observations
}
\description{
Actual function to find the site specific selection intensity that maximizes the likelihood of each tumor being mutated or not. Uses site and tumor specific mutation rates. Uses Brent 1 dimensional optimization technique.optimize_gamma
}
