% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_epistasis.R
\name{pairwise_variant_epistasis}
\alias{pairwise_variant_epistasis}
\title{Calculate SIs at variant level under pairwise epistasis model}
\usage{
pairwise_variant_epistasis(cesa, variant_pair)
}
\description{
Calculate SIs at variant level under pairwise epistasis model
}
\keyword{internal}
