% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_refset_name.R
\name{parse_refset_name}
\alias{parse_refset_name}
\title{Parse user-supplied refset name}
\usage{
parse_refset_name(refset_name)
}
\arguments{
\item{refset_name}{The name of a reference data package or directory.}
}
\description{
Parse user-supplied refset name
}
\keyword{internal}
