% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_signature_effects.R
\name{plot_signature_effects}
\alias{plot_signature_effects}
\title{Plot mutational source and effect attributions}
\usage{
plot_signature_effects(
  mutational_effects = NULL,
  signature_groupings = "auto",
  viridis_option = NULL,
  num_sig_groups = 7
)
}
\arguments{
\item{mutational_effects}{Output from \link{mutational_signature_effects}(). To compare
groups of samples, supply a named list with each element corresponding to output
from a separate run of mutational_signature_effects().}

\item{signature_groupings}{A data.table of signature names and descriptions; signatures with
  identical descriptions are grouped together. Only signatures present in the data get displayed.
  Setting to "auto" (the default) uses the table returned by \code{\link{cosmic_signature_info}()}, which only
  makes sense when using COSMIC signatures. A custom table should have columns "name",
  "short_name", and "description". Additional options:
  \itemize{
  \item To force a signature group to appear in the plot even if it
  has a low effect share, add a column called "prioritize" and set to TRUE where desired.
  \item To make a signature appear in its own group, make its description unique.
  \item Add a "color" column to manually specify colors for each group.
  }
Alternatively, setting \code{signature_groupings = "cannataro"} applies the same signature
groupings and color palette as
\href{https://academic.oup.com/mbe/article/39/5/msac084/6570859}{Cannataro et al. 2022}. You can use
Cannataro signature groupings with a different color palette by specifying \code{viridis_option}.}

\item{viridis_option}{A viridis color mapping, specified with a single letter ('A' to 'H'). By
default, map 'G' (mako) unless using Cannataro signature groupings.}

\item{num_sig_groups}{How many groups of signatures to display. Groups are ordered by their
highest effect shares, and the rest get lumped into an "other signatures" group.}
}
\description{
Compare the extent to which mutational signatures contribute mutations (mutational source share)
to the degree to which they contribute high-effect mutations (cancer effect share).
}
