% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_data_handling.R
\name{preload_ref_data}
\alias{preload_ref_data}
\title{preload_ref_data}
\usage{
preload_ref_data(data_dir)
}
\description{
Used when loading or creating a CESAnalysis to load reference into an environment for quick access
}
\keyword{internal}
