% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_ASCAT3_segments.R
\name{prep_ASCAT3_segments}
\alias{prep_ASCAT3_segments}
\title{Prep ASCAT3 copy segment data}
\usage{
prep_ASCAT3_segments(segments, refset)
}
\arguments{
\item{segments}{docs coming}

\item{refset}{reference data package (or custom reference data)}
}
\description{
Prep ASCAT3 copy segment data
}
