% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_to_maf_table.R
\name{read_vcf}
\alias{read_vcf}
\title{Internal VCF parser}
\usage{
read_vcf(vcf, sample_id, vcf_name = sample_id)
}
\arguments{
\item{vcf}{VCF filename or VCF-like data.table.}

\item{sample_id}{1-length sample identifier.}

\item{vcf_name}{1-length identifier used in some user messages.}
}
\value{
MAF-like data.table
}
\description{
Used by vcfs_to_maf_table()
}
\keyword{internal}
