% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_deconstructSigs.R
\name{run_deconstructSigs}
\alias{run_deconstructSigs}
\title{deconstructSigs wrapper}
\usage{
run_deconstructSigs(
  tumor_trinuc_counts,
  signatures_df,
  signatures_to_remove,
  artifact_signatures,
  tri.counts.method
)
}
\arguments{
\item{tumor_trinuc_counts}{one-row data.frame of trinuc variant counts (in deconstructSigs order) for one tumor}

\item{signatures_df}{data.frame of signatures (see COSMIC v3 signatures included with package for format)}

\item{signatures_to_remove}{names of signatures in signatures_df to keep out of deconstructSigs and assign zero weights}

\item{artifact_signatures}{names of signatures that should be treated as experimental (e.g., sequencing) artifacts
(these have weights calculated but are then normalized out when calculating true trinuc proportions)}

\item{tri.counts.method}{exome/genome trinucleotide content normalization argument to pass to deconstructSigs (see its docs)}
}
\description{
This function gets called internally during tumor-specific trinucleotide mutation rate calculation.
}
