% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_mutation_rates.R
\name{run_dndscv}
\alias{run_dndscv}
\title{Internal function to run dNdScv}
\usage{
run_dndscv(mutations, gene_list, cv, refdb, gr_genes, ...)
}
\description{
Internal function to run dNdScv
}
\keyword{internal}
