% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CESAnalysis.R
\name{sample_info}
\alias{sample_info}
\title{sample_info}
\usage{
sample_info(cesa = NULL)
}
\arguments{
\item{cesa}{CESAnalysis object}
}
\description{
returns a data.table with info on all samples in the CESAnalysis (at least, all samples with any valid mutations)
}
