% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CESAnalysis.R
\name{samples}
\alias{samples}
\title{View sample metadata}
\usage{
samples(cesa = NULL)
}
\arguments{
\item{cesa}{CESAnalysis object}
}
\description{
returns a data.table with info on all samples in the CESAnalysis (at least, all samples with any valid mutations)
}
