% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_samples.R
\name{samples_with}
\alias{samples_with}
\title{Find samples with specified variants}
\usage{
samples_with(cesa, any_of = NULL)
}
\arguments{
\item{cesa}{CESAnalysis}

\item{any_of}{Select samples with ANY of the given variant names/IDs, such as
c("8:142506482_C>G", "KRAS G12C"). When a gene has multiple transcripts in reference
data, you may wish to use full IDs, such as "KRAS_G12C_ENSP00000256078".}
}
\description{
A convenience function to identify samples with specific variants.
}
