% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CESAnalysis.R
\name{save_cesa}
\alias{save_cesa}
\title{Save a CESAnalysis in progress}
\usage{
save_cesa(cesa, file)
}
\arguments{
\item{cesa}{CESAnalysis to save}

\item{file}{filename to save to (must end in .rds)}
}
\description{
Saves a CESAnalysis to a file by calling using base R's saveRDS function. Also updates
run history for reproducibility. Files saved should be reloaded with \code{load_cesa()}.
}
\details{
Note that the genome reference data associated with a CESAnalysis (refset) is not
actually part of the CESAnalysis, so it is not saved here. (Saving this data with the
analysis would make file sizes too large.) When you reload the CESAnalysis, you can
re-associate the correct reference data.
}
