% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_samples.R
\name{select_samples}
\alias{select_samples}
\title{Retrieve validated subset of CESAnalysis samples table}
\usage{
select_samples(cesa = NULL, samples = character())
}
\arguments{
\item{cesa}{CESAnalysis}

\item{samples}{Vector of patient_ids, or data.table consisting
of rows from a CESAnalysis samples table. If empty, returns full sample table.}
}
\value{
data.table consisting of one or more rows from the CESAnalysis samples table.
}
\description{
Retrieve validated subset of CESAnalysis samples table
}
\keyword{internal}
