% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_intensity_calc_function.R
\name{selection_intensity_calculation}
\alias{selection_intensity_calculation}
\title{Selection intensity calculation (SNV)}
\usage{
selection_intensity_calculation(genes_for_analysis = "all", MAF_for_analysis,
  this.substitution = c("fake_gene", 41, "T"), trinuc.mutation_data,
  mut_rates, dndscv_siggenes, tumor.number = NULL,
  trinuc_count_matrix = NULL, output_from_mainMAF = NULL,
  sample_ID_column = "Unique_patient_identifier",
  gene_ID_column = "Gene_name", ref_column = "Reference_Allele",
  alt_column = "Tumor_allele", chr_column = "Chromosome",
  pos_column = "Start_Position", every_gene_verbose = F)
}
\arguments{
\item{genes_for_analysis}{String of genes, e.g. c("KRAS","EGFR","CTNNB1"), to calculate effect size.
If "all", the function calculates the effect size of all genes within the
data frame.}

\item{MAF_for_analysis}{Data frame in MAF format
(\url{https://docs.gdc.cancer.gov/Data/File_Formats/MAF_Format/})
with single nucleotide substitution data. Must contain columns with
gene names, chromosome, chromosomal position, reference allele, and mutant allele.}

\item{this.substitution}{Optional parameter to introduce a custom substitution
into the mutation rate calculation, e.g. like KRAS G12C here \url{https://www.nature.com/articles/s41388-017-0105-z}.}

\item{trinuc.mutation_data}{Data frame with trinucleotide mutation proportions for the tumor type.}

\item{mut_rates}{Vector with gene-level mutation rate data.}

\item{dndscv_siggenes}{dndscv data to be integrated into results data table.}

\item{tumor.number}{Number of tumor samples.
Useful if you split the analysis up into several independent runs. If \code{NULL}
then automatically sets to number of unique samples in \code{MAF_for_analysis}.}

\item{output_from_mainMAF}{the output from a previous selection run with all the
mutation and selection data, useful for bootstrapping because it speeds up the
analysis (no longer need to look up mutational context, etc.)}

\item{sample_ID_column}{column in MAF with sample ID data}

\item{gene_ID_column}{column in MAF with gene ID data}

\item{ref_column}{column in MAF with reference allele data}

\item{alt_column}{column in MAF with alternative allele data}

\item{chr_column}{column in MAF with chromosome data}

\item{pos_column}{column in MAF with chromosome nucleotide location data}
}
\description{
Function that calculates the selection intensity of single nucleotide
variants in cancer exome data.
}
