% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_results_converter.R
\name{selection_results_converter}
\alias{selection_results_converter}
\title{Selection results converter}
\usage{
selection_results_converter(results_input, subset_greater_than_freq = 1)
}
\arguments{
\item{results_input}{List of the selection results from get_gene_results}

\item{subset_greater_than_freq}{subsetting the data for only substitutions in more than this number of tumors}
}
\description{
Selection results converter
}
\examples{

}
