% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_gene_rates.R
\name{set_gene_rates}
\alias{set_gene_rates}
\title{Assign pre-calculated neutral gene mutation rates}
\usage{
set_gene_rates(
  cesa = NULL,
  gene_rates = NULL,
  sample_group = NULL,
  missing_genes_take_nearest = FALSE
)
}
\arguments{
\item{cesa}{CESAnalysis object}

\item{gene_rates}{A two-column data.table with genes in column 1 and rate in column 2}

\item{sample_group}{Character vector giving sample group(s) that the rates apply to. If
unspecified, the rates are assigned to all samples.}

\item{missing_genes_take_nearest}{Set to TRUE
to have each gene missing from gene_rates take the rate of the nearest non-missing gene.}
}
\description{
This function allows you to assign gene mutation rates calculated however you like to
samples in your CESAnalysis. Rates can be assigned to all samples or to samples
with the specified sample_group labels (see \code{?CESAnalysis}).
}
\details{
Provide rates in a gene_rates data.table with two columns: gene (character) and rate
(numeric on \code{[0, 1]}). Gene names must match the CESAnalysis reference data; to
see all genes in your CESAnalysis \code{cesa}, call
\code{unique(cesa$reference_data$gene_ranges$names)}. If you don't want to supply rates
for every gene, set \code{missing_genes_take_nearest = T} to have each missing gene
take the rate of the nearest non-missing gene.
}
