% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CESAnalysis.R
\name{set_refset_dir}
\alias{set_refset_dir}
\title{Set reference data directory}
\usage{
set_refset_dir(cesa, dir)
}
\arguments{
\item{cesa}{CESAnalysis}

\item{dir}{path to data directory}
}
\description{
When working with custom reference data or loading a previously saved CESAnalysis in a
new environment, use this function to reassociate the location of reference data with
the analysis. (If \code{load_cesa()} didn't give you a warning when loading your
analysis, you probably don't need to use this function.)
}
