% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_trinuc_rates.R
\name{set_trinuc_rates}
\alias{set_trinuc_rates}
\title{set_trinuc_rates}
\usage{
set_trinuc_rates(cesa, trinuc_proportion_matrix)
}
\arguments{
\item{cesa}{CESAnalysis object}

\item{trinuc_proportion_matrix}{rows=tumors, columns=relative mutation rates (see description)}
}
\description{
Directly assign trinucleotide-context-specific relative SNV mutation rates to 
all tumors in your CESAnalysis by supplying a matrix with this function.
This is for those who want to use their own methods instead of those
provided in trinucleotide_mutation_weights() and group_average_trinuc_rates().
There must be one row per tumor in your data set, with row names matching
sample identifiers. There must be 96 columns, with column names exactly matching
the deconstructSigs naming and order (run this function with incorrect column names, 
and the names you need to use will be printed). Since CES uses relative trinuc rates,
rows must sum to 1, with all values greater than 0.
}
