% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_lollipop.R
\name{si_lollipops}
\alias{si_lollipops}
\title{Plot selection intensities}
\usage{
si_lollipops(si_list, title = "My SIs", max_sites = 50)
}
\arguments{
\item{si_list}{Named list of tables, usually created by subsetting from selection
tables; e.g., `list(KRAS = cesa$selection[gene == "KRAS"], TP53 =
cesa$selection[gene == "TP53"])`}

\item{title}{plot title to pass to ggplot}

\item{max_sites}{maximum number of variant sites to include per lollipop;
if you try to include too many, you may have a challenge
getting it to look good}
}
\value{
ggplot object with lollipops
}
\description{
Lollipop plots can be used to show which variants have high
selection intensities within or across groups of samples.
}
\details{
To keep the lollipops readable, no more than \code{max_sites} variants are shown for
any group. At the same time, all SIs that are within the limits of the output plot are
depicted (which means that specification of \code{max_sites} creates an implicit SI
floor).
}
\examples{
\dontrun{
# Compare cancer subtypes from different analyses
si_lollipops(si_list = list(luad = luad_cesa$selection, lusc = lusc_cesa$selection))

# Compare results for two genes
tp53 = cesa$selection[gene == 'TP53']
egfr = cesa$selection[gene == 'EGFR']
si_lollipops(si_list = list(EGFR = egfr, TP53 = tp53))
}
}
