% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CESAnalysis.R
\name{snv_results}
\alias{snv_results}
\title{snv_results}
\usage{
snv_results(cesa = NULL)
}
\arguments{
\item{cesa}{CESAnalysis object}
}
\description{
returns a data table of SNV effect sizes generated with ces_snv
}
