% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_variant_si_objectives.R
\name{sswm_lik}
\alias{sswm_lik}
\title{sswm_lik}
\usage{
sswm_lik(rates_tumors_with, rates_tumors_without)
}
\arguments{
\item{rates_tumors_with}{vector of site-specific mutation rates for all tumors with variant}

\item{rates_tumors_without}{vector of site-specific mutation rates for all eligible tumors without variant}
}
\description{
Generates log-likelihood function of site-level selection with "strong selection, weak
mutation" assumption. All arguments to this likelihood function factory are
automatically supplied by \code{ces_variant()}.
}
