% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_variant_si_objectives.R
\name{sswm_sequential_lik}
\alias{sswm_sequential_lik}
\title{sswm_sequential_lik}
\usage{
sswm_sequential_lik(rates_tumors_with, rates_tumors_without, sample_index)
}
\arguments{
\item{rates_tumors_with}{vector of site-specific mutation rates for all tumors with variant}

\item{rates_tumors_without}{vector of site-specific mutation rates for all eligible tumors without variant}

\item{sample_index}{named numeric associating samples with progression states}
}
\description{
As in sswm_lik, selection intensities are calculated at variant sites under a "strong
selection, weak mutation" assumption. In this version, each sample is assigned to one
of an ordered set of disease progression states, and selection is assumed to vary
across states. For example, in a two-state local/metastatic model, each variant has two
independent selection intensities. Metastatic samples could have acquired the variant
while in their current state or at some earlier time, while the local state selection
intensity applied.
}
\details{
All arguments to this likelihood function factory are automatically supplied by
\code{ces_variant()}.
}
