% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_cosmic_signatures_to_remove.R
\name{suggest_cosmic_signatures_to_remove}
\alias{suggest_cosmic_signatures_to_remove}
\title{Identify mutational signatures to exclude from analysis}
\usage{
suggest_cosmic_signatures_to_remove(
  cancer_type = NULL,
  treatment_naive = NULL,
  quiet = FALSE
)
}
\arguments{
\item{cancer_type}{See chart on website for supported cancer type labels}

\item{treatment_naive}{give TRUE if samples were taken pre-treatment; FALSE or leave NULL otherwise}

\item{quiet}{(default false) for non-interactive use, suppress explanations and advice}
}
\value{
a vector of signatures to feed to signatures_to_remove
}
\description{
Get suggestions on signatures_to_remove for trinuc_mutation_rates for COSMIC signatures v3 and later.
For details, see \code{vignette("cosmic_cancer_type_note")}.
}
