% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trinuc_profile_function_with_weights.R
\name{trinuc_profile_function_with_weights}
\alias{trinuc_profile_function_with_weights}
\title{Trinucleotide profiler}
\usage{
trinuc_profile_function_with_weights(input_MAF,
  signature_choice = "signatures.cosmic",
  minimum_mutations_per_tumor = 50, save_figs = F,
  tumor.name = "[Tumor name here]",
  sample_ID_column = "Unique_patient_identifier",
  chr_column = "Chromosome", pos_column = "Start_Position",
  ref_column = "Reference_Allele", alt_column = "Tumor_allele")
}
\arguments{
\item{tumor.name}{If \code{save.figs = T} then you should have a character string
here for the save file and plot title.}

\item{input.MAF}{MAF with mutation data}

\item{signature.choice}{Either "signatures.cosmic" or "signatures.nature2013".}

\item{minimum.mutation.per.tumor}{Set to 50, as per Rosenthal et al. (see Details)}

\item{save.figs}{Boolean, if T saves figures (requires \code{ggplot2})
to "Figures/*tumor.name*.pdf"}

\item{sample.ID.column}{name of the column in \code{input.MAF} that contains unique
sample identifiers.}

\item{chr.column}{name of the column in \code{input.MAF} that contains chromosome
information.}

\item{pos.column}{name of the column in \code{input.MAF} that contains SNV position
information.}

\item{ref.column}{name of the column in \code{input.MAF} that contains reference
allele information.}

\item{alt.column}{name of the column in \code{input.MAF} that contains alternative
allele information.}
}
\description{
Creates a dataframe with the average trinucleotide product among tumors
for 96 trinucleotide contexts using
\code{\link[deconstructSigs]{whichSignatures}}. Also outputs the individual
signature weights vs tumors.
}
\details{
Uses the \code{deconstructSigs} package from Rosenthal, R., McGranahan, N., Herrero, J., Taylor, B. S., & Swanton, C. (2016). deconstructSigs: delineating mutational processes in single tumors distinguishes DNA repair deficiencies and patterns of carcinoma evolution. Genome Biology, 17(1), 31. https://doi.org/10.1186/s13059-016-0893-4
}
