% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trinucleotide_mutation_weights.R
\name{trinucleotide_mutation_weights}
\alias{trinucleotide_mutation_weights}
\title{Trinucleotide mutation weights}
\usage{
trinucleotide_mutation_weights(cesa, algorithm_choice = "weighted",
  remove_recurrent = TRUE,
  signature_choice = "signatures_cosmic_May2019",
  artifact_accounting = T, signatures_to_remove = c("SBS25", "SBS31",
  "SBS32", "SBS35"), mutation_count_rules = T)
}
\arguments{
\item{cesa}{# CESAnalysis object}

\item{algorithm_choice}{The choice of the algorithm used in determining trinucleotide weights.
Defaults to `weighted`, where all tumors with >= 50 substitutions have mutation
rates directly from deconstructSigs, and tumors with < 50 substitutions have that rate weighted
by the average of all tumors >= 50 substitutions relative to the substitution count < 50 in those tumors.
Other options currently include `all_average`, where all tumors have a mutation rate equal to the
average of all tumors >= 50 substitutions, `nearest_neighbor` where all tumors <50 substitutions
have the mutation rate of the tumor with >=50 substitutions closest to the mutational
distribution profile of the low substitution tumor as determined by a distance matrix, and
`all_calculated`, where all tumors have the mutation rate from the deconstructSigs output regardless of substitution number.}

\item{remove_recurrent}{If TRUE, removes recurrent variants prior to calculating trinucleotide signatures and weights. If FALSE, uses all variants.}

\item{signature_choice}{Either "signatures_cosmic_May2019" (default) or "signatures.cosmic" (COSMIC signatures v2 originally packaged with deconstructSigs).}

\item{artifact_accounting}{Accounts for the fact that the COSMIC v3 artifacts were detected, and renormalizes to see contribution of "true" sources of mutational flux.}

\item{signatures_to_remove}{Removes signatures from full potential signatures to minimize "signature bleeding", along the rationale proposed within the manuscript that originally calculated the v3 signature set doi: https://doi.org/10.1101/322859. Use `NULL` to keep all signatures. Signatures must correspond to the signature names in `signature_choice`.}

\item{mutation_count_rules}{T/F on whether to follow the mutation count rules outlined in https://doi.org/10.1101/322859, the manuscript reported the v3 COSMIC signature set.}
}
\description{
Trinucleotide mutation weights
}
\examples{



}
