% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumor_allele_adder.R
\name{tumor_allele_adder}
\alias{tumor_allele_adder}
\title{Tumor allele column addition}
\usage{
tumor_allele_adder(Reference_Allele, Tumor_Seq_Allele1, Tumor_Seq_Allele2)
}
\arguments{
\item{Reference_Allele:}{MAF-style "Reference_Allele" vector}

\item{Tumor_Seq_Allele1:}{MAF-style "Tumor_Seq_Allele1" vector}

\item{Tumor_Seq_Allele2:}{MAF-style "Tumor_Seq_Allele2" vector}
}
\description{
Several MAF files recently encountered had different representations of
the "tumor allele" . This function attempts to normalize all of the
different representations and return an analysis-ready vector of tumor alleles
}
