% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_tumor_addition.R
\name{unique_tumor_addition_function}
\alias{unique_tumor_addition_function}
\title{Unique tumor addition function}
\usage{
unique_tumor_addition_function(MAF_file,
  non_TCGA_characters_to_keep = "all", sum_stats = T, figures = F)
}
\arguments{
\item{MAF_file}{MAF file with "Tumor_Sample_Barcode" column}

\item{non_TCGA_characters_to_keep}{numeric number of characters to keep
in "Tumor_Sample_Barcode" column for sample names that
do not start with "TCGA". If non-numeric object then defaults to
all characters}

\item{sum_stats}{Boolean, T prints summary() of number of mutations per tumor}

\item{figures}{Boolean, T saves histogram of sum.stats and saves as "Figures/mutations_per_tumor.pdf"}
}
\description{
Adds a column called "Unique_patient_identifier" to your MAF file,
useful if TCGA codes have more data than sample and you want summary
statistics of the number of mutations per tumor.
}
