% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covered_regions_handling.R
\name{update_covered_in}
\alias{update_covered_in}
\title{update_covered_in}
\usage{
update_covered_in(cesa)
}
\arguments{
\item{cesa}{CESAnalysis}
}
\value{
CESAnalysis with regenerated covered-in annotations
}
\description{
Updates the covered_in annotation for all variants
to include all covered regions in the CESAnalysis
}
\keyword{internal}
