% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_id_handling.R
\name{validate_nt_change}
\alias{validate_nt_change}
\title{validate_nt_change}
\usage{
validate_nt_change(variant_ids, bsg)
}
\arguments{
\item{variant_ids}{character vector of variant_ids (SBS, DBS, or similar)}

\item{bsg}{BSgenome for getting reference sequence}
}
\description{
Ensures non-indel nucleotide changes are valid for the given genome.
Inputting indel IDs will give an error.
}
\keyword{internal}
