% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_optimizer_args.R
\name{validate_optimizer_args}
\alias{validate_optimizer_args}
\title{Check custom optimizer arguments}
\usage{
validate_optimizer_args(optimizer_args)
}
\arguments{
\item{optimizer_args}{List of arguments/values to pass to the optimizer.}
}
\description{
Check custom optimizer arguments
}
\keyword{internal}
