% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_id_handling.R
\name{validate_sbs_ids}
\alias{validate_sbs_ids}
\title{validate_sbs_ids}
\usage{
validate_sbs_ids(sbs_ids, bsg)
}
\arguments{
\item{sbs_ids}{character vector of sbs_ids}

\item{bsg}{BSgenome for getting reference sequence}
}
\description{
Ensures SBS IDs are valid for the given genome
}
\keyword{internal}
