% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_data_handling.R
\name{validate_signature_set}
\alias{validate_signature_set}
\title{validate_signature_set}
\usage{
validate_signature_set(signature_set)
}
\arguments{
\item{signature_set}{signature set list (see docs for format)}
}
\description{
Checks if a custom CES signature is properly formatted; stops with an error if not
}
