% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_counts.R
\name{variant_counts}
\alias{variant_counts}
\title{Assess variant prevalence and coverage}
\usage{
variant_counts(cesa, variant_ids = character(), by = character())
}
\arguments{
\item{cesa}{CESAnalysis}

\item{variant_ids}{variant names ("KRAS G12C") or full variant IDs. If left empty, uses
non-overlapping variants as returned by `select_variants()` with \code{min_freq = 1}.}

\item{by}{Optionally, a vector of one or more sample table columns. Variant prevalence
and coverage data will be broken down by the groups defined by unique combinations of
values in these columns.}
}
\description{
Determine variant prevalence (and how many samples have sequencing coverage) across your MAF data,
or within different groups of samples.
}
