% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_to_maf_table.R
\name{vcfs_to_maf_table}
\alias{vcfs_to_maf_table}
\title{Read a VCF into an MAF-like table}
\usage{
vcfs_to_maf_table(vcfs, sample_ids)
}
\arguments{
\item{vcfs}{Vector of VCF file paths, or a list of VCF-like data.tables, or a single data.table.}

\item{sample_ids}{Identifiers to populate Tumor_Sample_Barcode, one per VCF.}
}
\value{
A single data.table with MAF-style fields, suitable for use with cancereffectsizeR.
}
\description{
This function loads VCF files into MAF-like tables. A Tumor_Sample_Barcode column is
added, and the contents of the POS/REF/ALT fields are converted to match the style used
by MAF files for Start_Position/Reference_Allele/Tumor_Seq_Allele2. The VCF file
should represent high-confidence somatic variant calls for a single tumor sample.
}
