% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrongRef_dndscv_checker.R
\name{wrongRef_dndscv_checker}
\alias{wrongRef_dndscv_checker}
\title{\code{\link[dndscv]{dndscv}} wrong reference checker}
\usage{
wrongRef_dndscv_checker(mutations, gene_list = NULL, refdb = "hg19",
  sm = "192r_3w", kc = "cgc81", cv = "hg19",
  max_muts_per_gene_per_sample = 3, max_coding_muts_per_sample = 3000,
  use_indel_sites = T, min_indels = 5, maxcovs = 20,
  constrain_wnon_wspl = T, outp = 3)
}
\arguments{
\item{mutations}{Same as in \code{\link[dndscv]{dndscv}}. Table of
mutations (5 columns: sampleID, chr, pos, ref, alt). Only list independent events as mutations.}
}
\value{
matrix of provided mutations that do not match reference genome
}
\description{
Checks if reference base does not match reference within \code{\link[dndscv]{dndscv}} and tells you if there is a mismatch so you can remove. All inputs and most code same as in \code{\link[dndscv]{dndscv}}. See ?dndscv for more information.
}
