% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_handling.R
\name{add_variants}
\alias{add_variants}
\title{Add variant annotations}
\usage{
add_variants(target_cesa = NULL, snv_id = NULL, source_cesa = NULL)
}
\arguments{
\item{target_cesa}{CESAnalysis to receive variant annotations}

\item{snv_id}{character vector of CES-style snv_ids to validate, annotate (including
finding and adding associated amino-acid-change mutations), and add to the target
CESAnalysis}

\item{source_cesa}{source CESAnalysis whose snv_ids will be used to create new
annotations in the target CESAnalysis}
}
\description{
Add variant annotations
}
