% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_gene_maf.R
\name{annotate_gene_maf}
\alias{annotate_gene_maf}
\title{annotate_gene_maf}
\usage{
annotate_gene_maf(cesa)
}
\arguments{
\item{cesa}{CESAnalysis object}
}
\description{
Annotates CESAnaysis MAF data with gene information, keeping assignments consistent with dndscv when possible
}
