% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_variants.R
\name{annotate_variants}
\alias{annotate_variants}
\title{Annotate variants}
\usage{
annotate_variants(refset = NULL, variants = NULL)
}
\arguments{
\item{refset}{CES reference data set (e.g., from the ces.refset.hg19 data package)}

\item{variants}{MAF-like data.table of variants (e.g., as generated by preload_maf())}
}
\description{
Annotates CESAnaysis MAF data with reference genome and gene data; called by load_maf
}
\keyword{internal}
