% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_rate_calc.R
\name{baseline_mutation_rates}
\alias{baseline_mutation_rates}
\title{Baseline mutation rate calculation}
\usage{
baseline_mutation_rates(
  cesa,
  aac_ids = NULL,
  snv_ids = NULL,
  variant_ids = NULL,
  samples = NULL,
  cores = 1
)
}
\arguments{
\item{cesa}{CESAnalysis with gene mutation rates and tumor-specific trinucleotide-context-specific mutation rates already calculated}

\item{aac_ids}{vector of IDs for amino acid change variants}

\item{snv_ids}{vector of IDs for SNVs}

\item{variant_ids}{vector of mixed IDs (faster to use snv_ids and aac_ids for large jobs, if already known)}

\item{samples}{vector of sample IDs (Unique_Patient_Identifier) to include in mutation rate table (defaults to all samples)}

\item{cores}{number of cores to use for mutation processing (useful for large data sets or mutation lists)}
}
\value{
a data table of mutation rates with one column per variant, and a Unique_Patient_Identifier column identifying each row
}
\description{
Calculates neutral mutation rates at specific sites based on gene mutation rates and the relative
trinucleotide-context-specific SNV mutation rates of each sample
}
