% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_RefCDS.R
\name{build_RefCDS}
\alias{build_RefCDS}
\title{cancereffectsizeR's RefCDS builder}
\usage{
build_RefCDS(
  gtf,
  genome,
  output_by = "gene",
  cds_ranges_lack_stop_codons = T,
  cores = 1,
  numcode = 1
)
}
\arguments{
\item{gtf}{Path of a Gencode-GTF-formatted text file, or an equivalently formatted data
table. See details for required columns (features). It's possible to build a suitable
table using data pulled from biomaRt, but it's easier to use a GTF.}

\item{genome}{genome assembly name (e.g., "hg19"); an associated BSgenome object must be
available to load.}

\item{output_by}{"gene" (default) or "transcript", indicating whether the RefCDS output
should have one CDS annotation per gene or transcript.}

\item{cds_ranges_lack_stop_codons}{The CDS records in Gencode GTFs don't include the
stop codons in their genomic intervals. If your input table does include the stop 
codons within CDS records, set to FALSE.}

\item{cores}{how many cores to use for parallel computations}

\item{numcode}{(don't use) NCBI genetic code number; currently only code 1, the
standard genetic code, is supported}
}
\value{
A two-item list: RefCDS (which is itself a big list, with one entry per gene or
  transcript), and gr_genes, a GRanges object that defines the genomic intervals
  covered by each gene.
}
\description{
Based on the buildref function in Inigo Martincorena's package dNdScv, this function
takes in gene/transcript/CDS data and creates a dNdScv-style RefCDS object and
gr_genes, an associated GenomicRanges object also required to run dNdScv. For each gene
or transcript, only the longest complete CDS sequence is encoded in RefCDS output.
}
\details{
Required columns are seqnames, start, end, strand, gene_name, gene_id, protein_id, and
type. When output_by = "transcript", there must also be a transcript_id column. Only
rows that have type == "CDS" will be used. Strand should be "+" or "-".
}
