% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_baseline_mutation_rates.R
\name{calc_baseline_mutation_rates}
\alias{calc_baseline_mutation_rates}
\title{Runs functions supplying trinculeotide mutation weightings, mutation rate calculations, and MAF gene annotations}
\usage{
calc_baseline_mutation_rates(cesa = NULL, covariate_file = NULL,
  tumor_specific_rate_choice = F, trinuc_all_tumors = T,
  signature_choice = "signatures_cosmic_May2019",
  trinuc_algorithm_choice = "weighted", artifact_accounting = T,
  signatures_to_remove = c("SBS25", "SBS31", "SBS32", "SBS35"),
  mutation_count_rules = T, ...)
}
\arguments{
\item{cesa}{CESAnalysis object}

\item{covariate_file}{tissue-specific covariate file for dNdScv (gene-level mutation rate calculation)}

\item{trinuc_all_tumors}{Calculates trinucleotide signatures within all tumors (even those with < 50 variants)}

\item{signature_choice}{Either "signatures_cosmic_May2019" (default) or "signatures.cosmic" (COSMIC signatures v2 originally packaged with deconstructSigs).}

\item{signatures_to_remove}{Removes signatures from full potential signatures to minimize "signature bleeding", along the rationale proposed within the manuscript that originally calculated the v3 signature set doi: https://doi.org/10.1101/322859. Use `NULL` to keep all signatures. Signatures must correspond to the signature names in `signature_choice`.}

\item{mutation_count_rules}{T/F on whether to follow the mutation count rules outlined in https://doi.org/10.1101/322859, the manuscript reported the v3 COSMIC signature set.}
}
\description{
Runs functions supplying trinculeotide mutation weightings, mutation rate calculations, and MAF gene annotations
}
