% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_baseline_mutation_rates.R
\name{calc_baseline_mutation_rates}
\alias{calc_baseline_mutation_rates}
\title{calc_baseline_mutation_rates}
\usage{
calc_baseline_mutation_rates(
  cesa = NULL,
  covariate_file = NULL,
  cores = 1,
  signatures = "cosmic_v3"
)
}
\arguments{
\item{cesa}{CESAnalysis object}

\item{covariate_file}{tissue-specific covariate file for dNdScv (gene-level mutation rate calculation)}

\item{cores}{number of cores to use for calcualtions that can parallelized (requires parallel package)}

\item{signatures}{mutational signatures to which to attribute mutational processes in each tumor: "cosmic_v3" (default), 
"cosmic_v2" ,or a properly-formatted user-supplied signatures data.frame}
}
\description{
Wrapper to sequentially run the functions trinucleotide_mutation_weights, gene_level_mutation_rates, and
annotate_gene_maf. If you want to run these functions with advanced, non-default settings, see the
documentation for each function and run the functions individually instead of using this wrapper.
}
