% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_epistasis.R
\name{ces_epistasis}
\alias{ces_epistasis}
\title{Variant-level pairwise epistasis}
\usage{
ces_epistasis(cesa = NULL, variants = NULL, cores = 1, conf = NULL)
}
\arguments{
\item{cesa}{CESAnalysis}

\item{variants}{To test pairs of variants, supply a list where each element is a
2-length vector of CES-style variant IDs. Alternatively (and often more usefully),
supply a CompoundVariantSet (see \code{define_compound_variants}) to test all pairs
of compound variants in the set.}

\item{cores}{number of cores for parallel processing of gene pairs}

\item{conf}{confidence interval size from 0 to 1 (.95 -> 95%); NULL skips calculation,
reducing runtime.}
}
\value{
a data table with pairwise-epistatic selection intensities and variant
  frequencies for in tumors that have coverage at both variants in each pair
}
\description{
Calculate selection intensity under an assumption of pairwise epistasis between pairs of variants.
Returns a table, \strong{NOT} a CESAnalysis, so be careful not to overwrite your analysis.
}
