% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_epistasis.R
\name{ces_epistasis}
\alias{ces_epistasis}
\title{Variant-level pairwise epistasis}
\usage{
ces_epistasis(cesa = NULL, variants = NULL, cores = 1, conf = NULL)
}
\arguments{
\item{cesa}{CESAnalysis}

\item{variants}{list where each element is a character-type pair of variant IDs (either
amino-acid-change (coding) variant IDs or SNV IDs)}

\item{cores}{number of cores for parallel processing of gene pairs}

\item{conf}{confidence interval size from 0 to 1 (.95 -> 95%); NULL skips calculation,
significantly reducing runtime}
}
\value{
a data table with pairwise-epistatic selection intensities and variant
  frequencies for in tumors that have coverage at both variants in each pair
}
\description{
Calculate selection intensity under an assumption of pairwise epistasis between pairs of variants.
Return a table, \strong{NOT} a CESAnalysis, so be careful not to overwrite your analysis.
}
