% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_epistasis.R
\name{ces_epistasis}
\alias{ces_epistasis}
\title{Variant-level pairwise epistasis}
\usage{
ces_epistasis(cesa = NULL, variants = NULL, cores = 1)
}
\arguments{
\item{cesa}{CESAnalysis}

\item{variants}{list where each element is a character-type pair of variant IDs (either amino-acid-change (coding) variant IDs or SNV IDs)}

\item{cores}{number of cores for parallel processing of gene pairs}
}
\value{
a data.table with pairwise-epistatic selection intensities and variant frequency info
}
\description{
Calculate selection intensity under an assumption of pairwise epistasis between pairs of variants
}
