% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epistasis_wrappers.R
\name{ces_epistasis}
\alias{ces_epistasis}
\title{Variant-level pairwise epistasis}
\usage{
ces_epistasis(
  cesa = NULL,
  variants = NULL,
  samples = character(),
  run_name = "auto",
  cores = 1,
  model = "default",
  lik_args = list(),
  pval_calc_fn = NULL,
  optimizer_args = list(),
  conf = 0.95,
  return_fit = FALSE
)
}
\arguments{
\item{cesa}{CESAnalysis}

\item{variants}{To test pairs of variants, supply a list where each element is a
2-length vector of CES-style variant IDs. Alternatively (and often more usefully),
supply a VariantSetList (see \code{define_variant_sets()}) to test all pairs
of compound variants in the set.}

\item{samples}{Which samples to include in inference. Defaults to all samples.
Can be a vector of patient_ids, or a data.table containing rows from
the CESAnalysis sample table.}

\item{run_name}{Optionally, a name to identify the current run.}

\item{cores}{number of cores for parallel processing of variant pairs}

\item{model}{Set to "default"  to use built-in
model of epistatic selection, or supply a custom function factory (see details).}

\item{lik_args}{Extra arguments, given as a list, to pass to custom likelihood functions.}

\item{pval_calc_fn}{For use with custom models; optional. A function that takes an epistasis
model fit as input and returns p-values and other descriptives.}

\item{optimizer_args}{List of arguments to pass to the optimizer (bbmle::mle2).}

\item{conf}{confidence interval size from 0 to 1 (.95 -> 95\%); NULL skips calculation,
reducing runtime.}

\item{return_fit}{TRUE/FALSE (default FALSE): Embed epistatic model fits for each variant pair in
a "fit" attribute of the epistasis results table. Use \code{attr(my_results, 'fit')} to access
the list of fitted models.}
}
\value{
CESAnalysis with a table of epistatic inferences appended to list \code{[CESAnalysis]$epistasis}. Some column definitions:
  \itemize{
   \item variant_A, variant_B: Names for the two variants or merged sets of variants in each
   epistatic inference. For brevity in the case of merged variant sets, we say that a sample with
   any variant in variant set A "has variant A."
   \item ces_A0: Cancer effect (scaled selection coefficient) of variant A that acts in the absence of variant B.
   \item ces_B0: Cancer effect of variant B that acts in the absence of variant A.
   \item ces_A_on_B: Cancer effect of variant A that acts when a sample already has variant B.
   \item ces_B_on_A: Cancer effect of variant B that acts when a sample already has variant A.
   \item p_A_change: P-value of likelihood ratio test (LRT) that informs whether selection for
   variant A significantly changes after acquiring variant B. The LRT compares the likelihood of
   the full epistatic model to that of a reduced model in which ces_A0 and ces_A_on_B are set
   equal. The p-value is the probability, under the reduced model, of the likelihood ratio being
   greater than or equal to the ratio observed.
   \item p_B_change:  P-value of likelihood ratio test (LRT) that informs whether selection for
   variant B significantly changes after acquiring variant A. The LRT compares the likelihood of
   the full epistatic model to that of a reduced model in which ces_B0 and ces_B_on_A are set
   equal. The p-value is the probability, under the reduced model, of the likelihood ratio being
   greater than or equal to the ratio observed.
  \item p_epistasis: P-value of likelihood ratio test that informs whether the epistatic model
  better explains the mutation data than a non-epistatic model in which selection for mutations
  in each variant are independent of the mutation status of the other variant. Quite often, p_epistasis
  will suggest a significant epistatic effect even though p_A_change and p_B_change do not suggest
  significant changes in selection for either variant individually. This is because the degree of
  co-occurrence can often be explained equally well by a strong change in selection for either variant.
  \item expected_nAB_epistasis: The expected number of samples with both A and B mutated under the fitted epistatic model.
  Typically, this will be very close to the actual number of AB samples (nAB).
  \item expected_nAB_null: The expected number of samples with both A and B mutated under a no-epistasis model.
  \item AB_epistatic_ratio: The ratio \code{expected_nAB_epistasis/expected_nAB_null}. Useful to gauge the overall
  impact of epistatic interactions on the co-occurrence of variants A and B. Since the expectations take mutation rates into account,
  this ratio is a better indicator than the relative frequencies of A0, B0, AB, 00 in the data set.
  \item nA0, nB0, nAB, n00: Number of (included) samples with mutations in just A, just B, both A and B, and neither.
  \item ces_A_null, ces_B_null: Cancer effects of A and B when estimated independently; that is,
  effects under a no-epistasis model.
 }
}
\description{
Calculate selection intensity under an assumption of pairwise epistasis between pairs of variants.
VariantSetLists are supported.
}
