% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ces_gene_epistasis.R
\name{ces_gene_epistasis}
\alias{ces_gene_epistasis}
\title{Calculate gene-level selection intensity under an assumption of pairwise gene epistasis}
\usage{
ces_gene_epistasis(
  cesa = NULL,
  genes,
  cores = 1,
  full_gene_epistasis_lower_optim = 0.001,
  full_gene_epistasis_upper_optim = 1e+09,
  full_gene_epistasis_fnscale = -1e-16
)
}
\arguments{
\item{cesa}{CESAnalysis object}

\item{genes}{which genes to calculate effect sizes within; defaults to all genes with recurrent mutations in data set}

\item{cores}{number of cores to use}

\item{full_gene_epistasis_lower_optim}{optimx parameter}

\item{full_gene_epistasis_upper_optim}{optimx parameter}

\item{full_gene_epistasis_fnscale}{optimx parameter}
}
\value{
CESAnalysis object with selection results added for the chosen analysis
}
\description{
Calculate gene-level selection intensity under an assumption of pairwise gene epistasis
}
