% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_epistasis.R
\name{ces_gene_epistasis}
\alias{ces_gene_epistasis}
\title{Gene-level epistasis}
\usage{
ces_gene_epistasis(
  cesa = NULL,
  genes = character(),
  samples = character(),
  run_name = "auto",
  cores = 1,
  conf = NULL
)
}
\arguments{
\item{cesa}{CESAnalysis object}

\item{genes}{Vector of gene names; SIs will be calculated for all gene pairs.}

\item{samples}{Which samples to include in inference. Defaults to all samples. Can be a
vector of Unique_Patient_Identifiers, or a data.table containing rows from the
CESAnalysis sample table. NOTE: Even excluded samples are used to determine recurrent
variant sites. If this is a problem for you, then it's possible to get your desired
behavior using ces_epistasis().}

\item{run_name}{Optionally, a name to identify the current run.}

\item{cores}{number of cores for parallel processing of gene pairs}

\item{conf}{Confidence interval size from 0 to 1 (.95 -> 95\%). NULL skips calculation,
which reduces runtime.}
}
\value{
CESAnalysis with epistasis analysis results added
}
\description{
Calculate selection intensity under an assumption of pairwise gene-level epistasis.
More specifically, selection at the gene level is assumed to act through all recurrent
variants (variants that appear in just one sample are ignored), with selection
intensity of each gene assumed to vary based on mutational status of the other gene.
}
\details{
Only samples that have coverage at all recurrent sites in both genes can be included in
analysis since samples lacking full coverage may or may not have mutations at the
uncovered sites.
}
