% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_epistasis.R
\name{ces_gene_epistasis}
\alias{ces_gene_epistasis}
\title{Gene-level epistasis}
\usage{
ces_gene_epistasis(
  cesa = NULL,
  genes = character(),
  variants = NULL,
  samples = character(),
  run_name = "auto",
  cores = 1,
  conf = 0.95
)
}
\arguments{
\item{cesa}{CESAnalysis object}

\item{genes}{Vector of gene names; SIs will be calculated for all gene pairs.}

\item{variants}{Which variants to include in inference for each gene. Either
"recurrent" for all variants present in two or more samples (across all MAF data),
"nonsilent" for all coding variants except non-splice-site synonymous variants, or a
data.table containing all variants to include (as returned by
\code{select_variants()} or by subsetting\code{[CESAnalysis\]$variants}). For
noncoding variants with multiple gene annotations, the one listed in the "gene"
column is used. In the recurrent method, nearby noncoding variants may be included.}

\item{samples}{Which samples to include in inference. Defaults to all samples. Can be a
vector of Unique_Patient_Identifiers, or a data.table containing rows from the
CESAnalysis sample table.}

\item{run_name}{Optionally, a name to identify the current run.}

\item{cores}{number of cores for parallel processing of gene pairs}

\item{conf}{Confidence interval size from 0 to 1 (.95 -> 95\%). NULL skips calculation,
which may be helpful to reduce runtime when analyzing many gene pairs.}
}
\value{
CESAnalysis with epistasis analysis results added
}
\description{
Calculate selection intensity under an assumption of pairwise gene-level epistasis.
Selection at the gene level is assumed to act through all 
specified variants (see options), with the selection for each gene's variants allowed
to vary based on the mutational status of the other gene's variants.
}
\details{
Only samples that have coverage at all included sites in both genes can be included in the
inference since samples lacking full coverage may or may not have mutations at the
uncovered sites.
}
