% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_epistasis.R
\name{ces_gene_epistasis}
\alias{ces_gene_epistasis}
\title{Gene-level epistasis}
\usage{
ces_gene_epistasis(
  cesa = NULL,
  genes = character(),
  cores = 1,
  optimx_args = ces_epistasis_opm_args(),
  include_opm_report = FALSE
)
}
\arguments{
\item{cesa}{CESAnalysis object}

\item{genes}{vector of gene names; SIs will be calculated for all gene pairs}

\item{cores}{number of cores for parallel processing of gene pairs}

\item{optimx_args}{list of optimization parameters to feed into optimx::opm (see optimx docs); if you use this argument,
none of the CES default values will be used (opm defaults will be used for parameters you don't provide)}

\item{include_opm_report}{dev option to include full parameter optimization report with output (default false)}
}
\value{
a table of gene-level epistasis results
}
\description{
Calculate selection intensity under an assumption of pairwise gene-level epistasis.
More specifically, selection at the gene level is assumed to act through all recurrent
variants (variants that appear in just one sample are ignored), with selection
intensity of each gene assumed to vary based on mutational status of the other gene.
Returns a table of results, \strong{NOT} a CESAnalysis, so be careful not to overwrite
your CESAnalysis.
}
\details{
Only samples that have coverage at all recurrent sites can be included in analysis
since samples lacking full coverage may or may not have mutations at the uncovered
sites
}
