% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ces_gene_epistasis.R
\name{ces_gene_epistasis}
\alias{ces_gene_epistasis}
\title{Calculate gene-level selection intensity under an assumption of pairwise gene epistasis}
\usage{
ces_gene_epistasis(
  cesa = NULL,
  genes = character(),
  cores = 1,
  optimx_args = ces_gene_epistasis_opm_args(),
  return_all_opm_output = FALSE
)
}
\arguments{
\item{cesa}{CESAnalysis object}

\item{genes}{which genes to calculate effect sizes within}

\item{cores}{number of cores to use}

\item{optimx_args}{list of optimization parameters to feed into optimx::opm (see optimx docs); if you use this argument,
none of the CES default values will be used (opm defaults will be used for parameters you don't provide)}

\item{return_all_opm_output}{whether to include full parameter optimization report with output (default false)}
}
\value{
CESAnalysis object with selection results added for the chosen analysis
}
\description{
Calculate gene-level selection intensity under an assumption of pairwise gene epistasis
}
