% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ces_snv.R
\name{ces_snv}
\alias{ces_snv}
\title{Calculate selection intensity for single-nucleotide variants and amino acid changes}
\usage{
ces_snv(
  cesa = NULL,
  genes = "all",
  cores = 1,
  include_genes_without_recurrent_mutations = F,
  find_CI = T
)
}
\arguments{
\item{cesa}{CESAnalysis object}

\item{cores}{number of cores to use}

\item{include_genes_without_recurrent_mutations}{default false; will increase runtime and won't find anything interesting}

\item{gene}{which genes to calculate effect sizes within; defaults to all genes with recurrent mutations in data set}
}
\value{
CESAnalysis object with selection results added for the chosen analysis
}
\description{
Calculate selection intensity for single-nucleotide variants and amino acid changes
}
