% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ces_snv.R
\name{ces_snv}
\alias{ces_snv}
\title{Calculate selection intensity for single-nucleotide variants and amino acid changes}
\usage{
ces_snv(
  cesa = NULL,
  genes = NULL,
  min_freq = 2,
  variant_ids = NULL,
  cores = 1,
  conf = 0.95
)
}
\arguments{
\item{cesa}{CESAnalysis object}

\item{genes}{limit analysis to just these genes (otherwise, all genes, unless variants is specified)}

\item{min_freq}{default 2; setting to 0 or 1 can be useful for
comparing sample groups or analyzing SIs across genomic regions}

\item{variant_ids}{When specified, CES-style variant IDs of variants to include. If "genes" is also specified,
the union is taken. If you want to use this parameter to calculate SI
confidence intervals at sites with no MAF variants, set min_freq = 0}

\item{cores}{number of cores to use}

\item{conf}{selection intensity confidence interval width (NULL skips calculation, speeds runtime)}
}
\value{
CESAnalysis object with selection results added for the chosen analysis
}
\description{
Calculate selection intensity for single-nucleotide variants and amino acid changes
}
