% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ces_variant.R
\name{ces_variant}
\alias{ces_variant}
\title{Calculate selection intensity for single-nucleotide variants and amino acid changes}
\usage{
ces_variant(
  cesa = NULL,
  genes = NULL,
  min_freq = 2,
  variant_ids = NULL,
  lik_fn = "sswm",
  group_ordering = NULL,
  custom_lik_args = list(),
  cores = 1,
  conf = 0.95
)
}
\arguments{
\item{cesa}{CESAnalysis object}

\item{genes}{limit analysis to just these genes (otherwise, all genes, unless variants is specified)}

\item{min_freq}{default 2; setting to 0 or 1 can be useful for
comparing sample groups or analyzing SIs across genomic regions}

\item{variant_ids}{When specified, CES-style variant IDs of variants to include, or
short names like "KRAS G12C" (or KRAS_G12C). If "genes" is also specified, the union
is taken. If you want to use this parameter to calculate SI confidence intervals at
sites with no MAF variants, set min_freq = 0.}

\item{lik_fn}{"sswm" or "sswm_sequential" to use built-in models of selection, or supply a
custom function factory (see details).}

\item{group_ordering}{for models (like sswm_sequential) that assume different groups of
samples are from sequential tumor progression states, a vector or list giving the
ordering of CESAnalysis groups. Examples: c("Primary", "Metastatic") is a
two-state ordering. Lists allow groups to share progression states: 
list("A", c("B", "C"), "D") means that group A is first, groups B and C 
belong to the same intermediate state, and group D is last.}

\item{custom_lik_args}{Extra arguments, given as a list, to pass to custom likelihood
functions.}

\item{cores}{number of cores to use}

\item{conf}{selection intensity confidence interval width (NULL skips calculation, speeds runtime)}
}
\value{
CESAnalysis object with selection results added for the chosen analysis
}
\description{
Calculate selection intensity for single-nucleotide variants and amino acid changes
}
