% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CESAnalysis.R
\name{CESAnalysis}
\alias{CESAnalysis}
\title{Create a cancereffectsizeR analysis}
\usage{
CESAnalysis(ref_set = "ces_hg19_v1", progression_order = NULL)
}
\arguments{
\item{ref_set}{reference data to use (e.g. "ces_hg19_v1")}

\item{progression_order}{evolutionary order of tumor stage progression (e.g. c("Primary", "Metastatic"))}
}
\value{
CESAnalysis object
}
\description{
Creates a CESAnalysis object, the central data structure of cancereffectsizeR.
}
