% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CESAnalysis.R
\name{CESAnalysis}
\alias{CESAnalysis}
\title{Create a cancereffectsizeR analysis}
\usage{
CESAnalysis(genome = NULL, progression_order = NULL)
}
\arguments{
\item{genome}{genome assembly name (e.g., "hg19"), if using built-in reference data}

\item{progression_order}{evolutionary order of tumor stage progression (e.g. c("Primary", "Metastatic"))}
}
\value{
CESAnalysis object
}
\description{
Creates a CESAnalysis object, the central data structure of cancereffectsizeR
}
