% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CESAnalysis.R
\name{CESAnalysis}
\alias{CESAnalysis}
\title{Create a cancereffectsizeR analysis}
\usage{
CESAnalysis(refset = NULL)
}
\arguments{
\item{refset}{Name of reference data set (refset) to use; run \code{list_ces_refsets()} for
available refsets. Alternatively, the path to a custom reference data directory.}
}
\value{
CESAnalysis object
}
\description{
Creates a CESAnalysis, the central data structure of cancereffectsizeR.
}
