% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CESAnalysis.R
\name{CESAnalysis}
\alias{CESAnalysis}
\title{Create a cancereffectsizeR analysis}
\usage{
CESAnalysis(ref_set = "ces_hg19_v1", sample_groups = NULL)
}
\arguments{
\item{ref_set}{reference data to use; run \code{list_ces_ref_sets()} for available ref sets.}

\item{sample_groups}{Optionally, supply labels identifying different groups of samples.
To be able to perform analyses that require ordered groups of samples, give the
labels in proper order: e.g., c("Primary", "Metastatic"). If you will not be doing
such analyses, ordering doesn't matter.}
}
\value{
CESAnalysis object
}
\description{
Creates a CESAnalysis object, the central data structure of cancereffectsizeR.
}
