% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CESAnalysis.R
\name{CESAnalysis}
\alias{CESAnalysis}
\title{CESAnalysis}
\usage{
CESAnalysis(genome = NULL, progression_order = NULL)
}
\arguments{
\item{genome}{Genome build of MAF data (currently, just hg19 supported)}

\item{progression_order}{evolutionary order of tumor stage progression (e.g. c("Primary", "Metastatic"))}
}
\value{
CESAnalysis object
}
\description{
Creates a CESAnalysis object (the data structure used by cancereffectsizeR)
}
