% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CESAnalysis.R
\name{CESAnalysis}
\alias{CESAnalysis}
\title{Create a cancereffectsizeR analysis}
\usage{
CESAnalysis(ref_set = "ces_hg19_v1", sample_groups = NULL)
}
\arguments{
\item{ref_set}{name of reference data set to use; run \code{list_ces_ref_sets()} for
available ref sets. Alternatively, the path to a custom reference data directory.}

\item{sample_groups}{Optionally, supply labels identifying different groups of samples.
Each designated group of samples can be run independently through functions like
\code{trinuc_mutation_rates()} and \code{gene_mutation_rates()}, and some selection
models (such as sswm_sequential) require multiple sample groups.}
}
\value{
CESAnalysis object
}
\description{
Creates a CESAnalysis object, the central data structure of cancereffectsizeR.
}
