% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CESAnalysis.R
\name{CESAnalysis}
\alias{CESAnalysis}
\title{Create a cancereffectsizeR analysis}
\usage{
CESAnalysis(refset = NULL, sample_groups = NULL)
}
\arguments{
\item{refset}{name of reference data set (refset) to use; run \code{list_ces_refsets()} for
available refsets. Alternatively, the path to a custom reference data directory.}

\item{sample_groups}{(Deprecated; no longer necessary.) Optionally, supply labels
identifying different groups of samples. Each designated group of samples can be run
independently through functions like \code{trinuc_mutation_rates()} and
\code{gene_mutation_rates()}, and some selection models (such as sswm_sequential)
require multiple sample groups.}
}
\value{
CESAnalysis object
}
\description{
Creates a CESAnalysis, the central data structure of cancereffectsizeR.
}
