% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_finder.R
\name{CI_finder}
\alias{CI_finder}
\title{MLE CI finder}
\usage{
CI_finder(gamma_max, all_tumors, MAF_input, progressions, gene, variant,
  specific_mut_rates, log_units_down = 10.828/2, threshold = 0.01)
}
\arguments{
\item{gamma_max}{selection intensity that maximizes the objective function}

\item{all_tumors}{A list of all the tumors we are calculating the likelihood across}

\item{MAF_input}{A data frame that includes columns "Unique_Patient_Identifier", "Gene_name", and "unique_variant_ID_AA"}

\item{progressions}{CESProgressions object}

\item{gene}{The gene we want to look at}

\item{variant}{The variant we want to look at}

\item{specific_mut_rates}{A matrix of site and tumor specific mutation rates where the rows correspond to tumors and the columns to variants (produced by mutation_rate_calc)}

\item{log_units_down}{log units below maximum that correspond to 99.9 percent confidence}

\item{threshold}{tolerance threshold for the uniroot function}
}
\description{
MLE CI finder
}
