% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CESAnalysis.R
\name{clean_granges_for_cesa}
\alias{clean_granges_for_cesa}
\title{clean_granges_for_cesa}
\usage{
clean_granges_for_cesa(cesa = NULL, gr = NULL, padding = 0, refset_env = NULL)
}
\arguments{
\item{cesa}{CESAnalysis}

\item{gr}{GRanges object}

\item{padding}{How many bases to expand start and end of each position}
}
\description{
Tries to format an input GRanges object to be compatible with a CESAnalysis reference
genome. Optionally, also applies padding to start and end positions of ranges, stopping
at chromosome ends. Either stops with an error or returns a clean granges object.
}
\keyword{internal}
