% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_refset.R
\name{create_refset}
\alias{create_refset}
\title{Create a custom refset}
\usage{
create_refset(
  output_dir,
  refcds_output,
  species_name,
  genome_build_name,
  BSgenome_name,
  supported_chr = c(1:22, "X", "Y"),
  default_exome_bed = NULL,
  exome_interval_padding = 0
)
}
\arguments{
\item{output_dir}{Name/path of an existing, writeable output directory where all data
will be saved. The name of this directory will serve as the name of the custom refset.}

\item{refcds_output}{Transcript information in the two-item list (consisting of RefCDS
and gr_genes) that is output by \code{build_RefCDS}.}

\item{species_name}{Name of the species, primarily for display (e.g., "human").}

\item{genome_build_name}{Name of the genome build, primarily for display (e.g., "hg19").}

\item{BSgenome_name}{The name of the BSgenome package to use (e.g., "hg19"); will used by
cancereffectsizeR to load the reference genome via BSgenome::getBSgenome().}

\item{supported_chr}{Character vector of supported chromosomes. Note that
cancereffectsizeR uses NCBI-style chromosome names, which means no chr prefixes ("X",
not "chrX"). Mitochondrial contigs shouldn't be included since they would require
special handling that hasn't been implemented.}

\item{default_exome_bed}{A BED file, typically acquired from some exome capture kit
documentation, that defines default exome capture intervals for the genome. When users run
cancereffectsizeR without specifying the genomic intervals that are covered by their
sequencing data, these intervals inform some calculations. They can also be used to
exclude non-covered variant calls. Including a default exome is optional but
recommended, and users can always choose to use their own coverage definitions
instead.}

\item{exome_interval_padding}{Number of bases to pad start/end of each covered
interval, to allow for some variants to be called just outside of targeted regions,
where there still may be pretty good sequencing coverage.}
}
\description{
Use this function to create and save a directory of custom reference data that can be
used with cancereffectsizeR instead of supplied refsets like \code{ces.refset.hg19}. All
arguments are required except default_exome/exome_interval_padding, which are recommended.
}
\details{
To run this function, you'll need to have output from \code{build_RefCDS()}.
}
