% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconstructSigs_input_preprocess.R
\name{deconstructSigs_input_preprocess}
\alias{deconstructSigs_input_preprocess}
\title{deconstructSigs_input_preprocess}
\usage{
deconstructSigs_input_preprocess(MAF,
  sample_ID_column = "Unique_patient_identifier",
  chr_column = "Chromosome", pos_column = "Start_Position",
  ref_column = "Reference_Allele", alt_column = "Tumor_allele")
}
\arguments{
\item{alt_column}{}
}
\description{
Gathers all SNV and removes recurrent variants
}
