% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_handling.R
\name{.add_covered_regions}
\alias{.add_covered_regions}
\title{.add_coverage_gr}
\usage{
.add_covered_regions(
  cesa,
  coverage_type,
  covered_regions,
  covered_regions_name,
  covered_regions_padding,
  update_anno
)
}
\description{
.add_coverage_gr
}
