% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_counts.R
\name{.variant_counts}
\alias{.variant_counts}
\title{Internal variant prevalence and coverage calculation}
\usage{
.variant_counts(
  cesa,
  samples,
  sbs_from_aac,
  noncoding_sbs_id,
  by_cols = character()
)
}
\arguments{
\item{cesa}{CESAnalysis}

\item{samples}{validated samples table}

\item{sbs_from_aac}{data.table with columns aac_id, sbs_id (validated and with annotations in CESAnalysis)}

\item{noncoding_sbs_id}{vector of sbs_ids to treat as noncoding variants}

\item{by_cols}{validated column names from sample table that are suitable to use for counting by.}
}
\description{
Called by variant_counts() (and select_variants()) with validated inputs.
}
\keyword{internal}
