% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size_calculation.R
\name{effect_size_SNV}
\alias{effect_size_SNV}
\title{SNV effect size calculation function}
\usage{
effect_size_SNV(covariate_file = NULL, ceslist = NULL,
  genes_for_effect_size_analysis = "all", cores = 1,
  tumor_specific_rate_choice = F, trinuc_all_tumors = T,
  subset_col = NULL, subset_levels_order = NULL,
  epistasis_top_prev_number = NULL, epistasis_gene_level = F,
  full_gene_epistasis_lower_optim = 0.001,
  full_gene_epistasis_upper_optim = 1e+09,
  full_gene_epistasis_fnscale = -1e-16,
  q_threshold_for_gene_level = 0.1,
  trinuc_algorithm_choice = "weighted")
}
\arguments{
\item{covariate_file}{Either NULL and uses the \code{dndscv}
default covariates, or one of these:  "bladder_pca"  "breast_pca"
"cesc_pca" "colon_pca" "esca_pca" "gbm_pca" "hnsc_pca" "kidney_pca" "lihc_pca" "lung_pca" "ov_pca" "pancreas_pca" "prostate_pca" "rectum_pca" "skin_pca"  "stomach_pca"  "thca_pca" "ucec_pca"}

\item{ceslist}{a CESList object}

\item{genes_for_effect_size_analysis}{genes to calculate effect sizes within. If unspecified, defaults to "all".}

\item{cores}{number of cores to use}

\item{tumor_specific_rate_choice}{weights tumor-specific rates by their relative proportional substitution count (not recommended)}

\item{trinuc_all_tumors}{Calculates trinucleotide signatures within all tumors (even those with < 50 variants)}

\item{subset_col}{column in MAF with subset data (e.g., column contains data like "Primary" and "Metastatic" in each row)}

\item{subset_levels_order}{evolutionary order of events in subset_col. (e.g. c("Primary", "Metastatic")}

\item{MAF}{MAF file with substitution data}
}
\description{
Function that determines mutation rate at the gene- and trinucleotide- level
and then calculates cancer effect size.
Works best on datasets with a large number of tumors, each with a large
number of SNV (The \code{deconstructSigs} package, used here to calculate
trinucleotide signatues, suggests at least 50 SNV per tumor used in the
signature calculation). Requres data to be in hg coordinates, you can use
the \code{hg_converter()} function provided with this package to convert
from one coordinate system to another.
}
