% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size_SNV.R
\name{effect_size_SNV}
\alias{effect_size_SNV}
\title{Calculate SNV selection intensity}
\usage{
effect_size_SNV(cesa = NULL, genes = "all", analysis = c("SNV",
  "gene-level-epistasis", "top-epistasis"),
  epistasis_top_prev_number = NULL, cores = 1,
  ignore_progression_stages = F,
  include_genes_without_recurrent_mutations = F,
  full_gene_epistasis_lower_optim = 0.001,
  full_gene_epistasis_upper_optim = 1e+09,
  full_gene_epistasis_fnscale = -1e-16,
  q_threshold_for_gene_level = 0.1, find_CI = T)
}
\arguments{
\item{cesa}{CESAnalysis object}

\item{analysis}{choose SNV, gene-level-epistasis, or top-epistasis}

\item{cores}{number of cores to use}

\item{ignore_progression_stages}{don't calculate stage-specific selection intensities even if stage data is present}

\item{include_genes_without_recurrent_mutations}{default false; will greatly slow runtime and won't find anything interesting}

\item{gene}{which genes to calculate effect sizes within; defaults to all genes with recurrent mutations in data set}
}
\value{
CESAnalysis object with selection results added for the chosen analysis
}
\description{
Calculate SNV selection intensity
}
