% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size_calculation.R
\name{effect_size_SNV}
\alias{effect_size_SNV}
\title{SNV effect size calculation function}
\usage{
effect_size_SNV(MAF_file, covariate_file = NULL, output_from_mainMAF = NULL,
  sample_ID_column = "Unique_patient_identifier", chr_column = "Chromosome",
  pos_column = "Start_Position", ref_column = "Reference_Allele",
  alt_column = "Tumor_allele", genes_for_effect_size_analysis = "all")
}
\arguments{
\item{MAF_file}{MAF file with substitution data}

\item{covariate_file}{Either NULL and uses the \code{dndscv}
default covariates, or one of these:  "bladder_pca"  "breast_pca"
"cesc_pca" "colon_pca" "esca_pca" "gbm_pca" "hnsc_pca" "kidney_pca" "lihc_pca" "lung_pca" "ov_pca" "pancreas_pca" "prostate_pca" "rectum_pca" "skin_pca"  "stomach_pca"  "thca_pca" "ucec_pca"}

\item{output_from_mainMAF}{the output from a previous selection run with all the
mutation and selection data, useful for bootstrapping because it speeds up the
analysis (no longer need to look up mutational context, etc.)}

\item{sample_ID_column}{column in MAF with sample ID data}

\item{chr_column}{column in MAF with chromosome data}

\item{pos_column}{column in MAF with chromosome nucleotide location data}

\item{ref_column}{column in MAF with reference allele data}

\item{alt_column}{column in MAF with alternative allele data}

\item{genes_for_effect_size_analysis}{genes to calculate effect sizes within.}
}
\description{
Function that determines mutation rate at the gene- and trinucleotide- level
and then calculates cancer effect size.
Works best on datasets with a large number of tumors, each with a large
number of SNV (The \code{deconstructSigs} package, used here to calculate
trinucleotide signatues, suggests at least 50 SNV per tumor used in the
signature calculation). Requres data to be in hg coordinates, you can use
the \code{hg_converter()} function provided with this package to convert
from one coordinate system to another.
}
