% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_level_mutation_rates.R
\name{gene_level_mutation_rates}
\alias{gene_level_mutation_rates}
\title{Gene-level mutation rate function}
\usage{
gene_level_mutation_rates(dndscv_output, RefCDS_object)
}
\arguments{
\item{dndscv_output}{output from dndscv}

\item{RefCDS_object}{object in the style of 'RefCDS' within the dndscv package}
}
\description{
This function calculates the maximum likelihood estimate for the expected
number of synonymous mutations in a gene under the dNdScv model, as described within
Martincorena, I., Raine, K. M., Gerstung, M., Dawson, K. J., Haase, K., Van Loo, P., … Campbell, P. J.
(2017). Universal Patterns of Selection in Cancer and Somatic Tissues. Cell. https://doi.org/10.1016/j.cell.2017.09.042
and the associated R package dndscv .
}
\examples{


}
