% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_level_mutation_rates.R
\name{gene_level_mutation_rates}
\alias{gene_level_mutation_rates}
\title{Gene-level mutation rate function}
\usage{
gene_level_mutation_rates(
  cesa,
  covariate_file = NULL,
  save_all_dndscv_output = FALSE
)
}
\arguments{
\item{cesa}{CESAnalysis object}

\item{covariate_file}{Either NULL and usefs dNdScv default covariates, or one of these:  "bladder_pca"  "breast_pca"
"cesc_pca" "colon_pca" "esca_pca" "gbm_pca" "hnsc_pca" "kidney_pca" "lihc_pca" "lung_pca" "ov_pca" "pancreas_pca" "prostate_pca" "rectum_pca" "skin_pca"  "stomach_pca"  "thca_pca" "ucec_pca"}

\item{save_all_dndscv_output}{default false; when true, saves all dndscv output, not just what's needed by CES}
}
\value{
CESAnalysis object with
}
\description{
This function calculates the maximum likelihood estimate for the expected
number of synonymous mutations in a gene under the dNdScv model, as described within
Martincorena, I., Raine, K. M., Gerstung, M., Dawson, K. J., Haase, K., Van Loo, P., … Campbell, P. J.
(2017). Universal Patterns of Selection in Cancer and Somatic Tissues. Cell. https://doi.org/10.1016/j.cell.2017.09.042
and the associated R package dndscv .
}
