% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_mutation_rates.R
\name{gene_mutation_rates}
\alias{gene_mutation_rates}
\title{Use dNdScv with tissue-specific covariates to calculate gene-level mutation rates}
\usage{
gene_mutation_rates(cesa, covariates = NULL, save_all_dndscv_output = FALSE)
}
\arguments{
\item{cesa}{CESAnalysis object}

\item{covariates}{Name of gene mutation rate covariates to use (run
list_ces_covariates() to see choices). For hg19 only, you can also use "default" for
dNdScv's non-tissue-specific covariates. If no covariates are available, set NULL to
run without.}

\item{save_all_dndscv_output}{default false; when true, saves all dndscv output, not
just what's needed by CES (will make object size very large)}
}
\value{
CESAnalysis object with gene-level mutation rates calculated
}
\description{
This function calculates gene-level neutral mutation rates based on counts
of nonsynonymous and synonymous mutations per gene under the dNdScv package's model, 
as described in Martincorena et al. (https://doi.org/10.1016/j.cell.2017.09.042).
}
