% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_data_handling.R
\name{get_ces_signature_set}
\alias{get_ces_signature_set}
\title{get_ces_signature_set}
\usage{
get_ces_signature_set(refset, name)
}
\arguments{
\item{refset}{name of refset (if using a custom refset, it must be loaded into a CESAnalysis already)}

\item{name}{name of signature set}
}
\description{
For a given CES reference data collection and signature set name, returns
cancereffectsizeR's internal data for the signature set in a three-item list: 
the signature set name, a data table of signature metadata, and a signature 
definition data frame
}
