% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ces_snv.R
\name{get_gene_results}
\alias{get_gene_results}
\title{Single-stage SNV effect size analysis (gets called by ces_snv)}
\usage{
get_gene_results(gene, cesa, all_tumors, find_CI)
}
\description{
Single-stage SNV effect size analysis (gets called by ces_snv)
}
