% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size_SNV.R
\name{get_gene_results}
\alias{get_gene_results}
\title{Single-stage SNV effect size analysis (gets called by effect_size_SNV)}
\usage{
get_gene_results(gene_to_analyze, cesa, gene_mafs, gene_trinuc_comp,
  tumor_specific_rate_choice, all_tumors, find_CI)
}
\description{
Single-stage SNV effect size analysis (gets called by effect_size_SNV)
}
