% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_data_handling.R
\name{get_genome_data_directory}
\alias{get_genome_data_directory}
\title{Get the path to the reference data directory of a CESAnalysis}
\usage{
get_genome_data_directory(genome)
}
\description{
takes in a genome name (e.g., "hg38"), searches for an associated data directory, and returns the path
}
