% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_data_handling.R
\name{get_genome_dirs}
\alias{get_genome_dirs}
\title{get_genome_dirs
returns a character vector mapping genome names to their data directories}
\usage{
get_genome_dirs()
}
\arguments{
\item{full_paths}{default TRUE; returns full paths to genome data instead of just genome names}
}
\description{
get_genome_dirs
returns a character vector mapping genome names to their data directories
}
