% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genomic_windows.R
\name{get_genomic_windows}
\alias{get_genomic_windows}
\title{Split genome into same-width ranges.}
\usage{
get_genomic_windows(
  refset,
  window_size = 1e+05,
  exclude_ENCODE_blacklist = TRUE,
  exclude_Y = TRUE,
  annotate_intervals = FALSE
)
}
\arguments{
\item{window_size}{Integer size of windows (default 100,000 bp). This function will be slow if
window_size is unexpectedly small.}

\item{exclude_ENCODE_blacklist}{TRUE/FALSE on filtering out windows in which at least 50% of
bases are within ENCODE blacklist regions (default TRUE).}

\item{exclude_Y}{TRUE/FALSE on whether to exclude chrY from output ranges; default TRUE.}

\item{annotate_intervals}{Add some gene-based annotation columns (default FALSE).}
}
\description{
Split genome into same-width ranges.
}
