% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CESAnalysis.R
\name{get_signature_weights}
\alias{get_signature_weights}
\title{Get table of signature attributions}
\usage{
get_signature_weights(cesa = NULL, raw = F, artifacts_zeroed = NULL)
}
\arguments{
\item{cesa}{CESAnalysis object}

\item{raw}{Default FALSE. When TRUE, return raw signature attributions as found by the
signature extraction tool. Format may vary by tool.}

\item{artifacts_zeroed}{Deprecated.}
}
\value{
A data.table of signature attributions for each sample. By default, these are
  estimated relative weights of biologically-associated signatures (i.e., non-artifact
  signatures) that sum to 1.
}
\description{
View SBS signature attributions associated with CESAnalysis samples.
}
\details{
Use raw = TRUE to get signature attributions as produced by the signature extraction
tool (or as provided by the user with set_signature_weights()), without any of the
adjustments that are made by cancereffectsizeR's trinuc_mutation_rates().
}
