% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CESAnalysis.R
\name{get_trinuc_rates}
\alias{get_trinuc_rates}
\title{Get expected relative trinucleotide-specific SNV mutation rates}
\usage{
get_trinuc_rates(cesa = NULL)
}
\arguments{
\item{cesa}{CESAnalysis object}
}
\description{
Get expected relative trinucleotide-specific SNV mutation rates
}
