% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_data_handling.R
\name{list_genomes}
\alias{list_genomes}
\title{list_genomes
prints names of  all genome builds that are ready for use with cancereffectsizeR}
\usage{
list_genomes()
}
\description{
list_genomes
prints names of  all genome builds that are ready for use with cancereffectsizeR
}
