% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_maf.R
\name{load_maf}
\alias{load_maf}
\title{load_maf}
\usage{
load_maf(cesa = NULL, maf = NULL,
  sample_col = "Tumor_Sample_Barcode", chr_col = "Chromosome",
  start_col = "Start_Position", ref_col = "Reference_Allele",
  tumor_allele_col = "guess", genome_build = "hg19",
  covered_regions = "exome", progression_col = NULL)
}
\arguments{
\item{cesa}{the CESAnalysis object to load the data into}

\item{maf}{Path of tab-delimited text file in MAF format, or an MAF in data.frame format}

\item{sample_col}{column name with sample ID data}

\item{chr_col}{column name with chromosome data}

\item{start_col}{column name with start position ( currently only analyze SNVs, so no end position needed)}

\item{ref_col}{column name with reference allele data}

\item{tumor_allele_col}{column name with alternate allele data (defaults to "guess", which examines ref_col,
"Tumor_Seq_Allele1", and "Tumor_Seq_Allele2" and determines from there)}

\item{genome_build}{genome build associated with data (must match build of CESAnalysis; currently just hg19 supported)}

\item{progression_col}{column in MAF with subset data (e.g., column contains data like "Primary" and "Metastatic" in each row)}

\item{coverged_regions}{for panel sequencing data, a BED file giving intervals covered by the panel}
}
\value{
CESAnalysis object with somewhat cleaned-up MAF data added
}
\description{
Load MAF-formatted mutation data into a CESAnalysis object (the core cancereffectsizeR object)
}
