% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_variant_si_objectives.R
\name{ml_objective}
\alias{ml_objective}
\title{ml_objective}
\usage{
ml_objective(
  tumor_stages,
  tumors_with_variant,
  tumors_without_gene_mutated,
  baseline_rates
)
}
\arguments{
\item{tumor_stages}{an environment with keys = tumor names, values = stage of tumor}

\item{tumors_with_variant}{vector of (eligible) tumors with the variant}

\item{tumors_without_gene_mutated}{vector of (eligible) tumors without any mutation in the gene of the variant}

\item{baseline_rates}{a named vector of baseline mutation rates for each eligible tumor}
}
\value{
loglikelihood value
}
\description{
Log-likelihood function under a model of site-specific selection intensities at variant sites
}
\keyword{internal}
