% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objective.R
\name{ml_objective}
\alias{ml_objective}
\title{ml_objective}
\usage{
ml_objective(gamma, MAF_input, all_tumors, gene, variant,
  specific_mut_rates)
}
\arguments{
\item{specific_mut_rates}{}
}
\description{
ml_objective
}
