% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objective_epistasis_full_gene.R
\name{ml_objective_epistasis_full_gene}
\alias{ml_objective_epistasis_full_gene}
\title{ml_objective for epistasis calculation at the gene level}
\usage{
ml_objective_epistasis_full_gene(
  par,
  with_just_1,
  with_just_2,
  with_both,
  with_neither
)
}
\arguments{
\item{par}{A selection intensity at which to calculate the likelihood}

\item{with_just_1}{vector of tumors with mutation in just the first gene}

\item{with_just_2}{vector of tumors with mutation in just the second gene}

\item{with_both}{vector of tumors with mutations in both genes}

\item{with_neither}{vector of tumors with no mutations in either gene}
}
\value{
A log likelihood value
}
\description{
Objective function that we will be optimizing (minimizing the negative of the log likelihood, so maximizing the log likelihood) in order to find the site specific selection intensity that maximizes the likelihood of each tumor having a mutation or not, where the mutation rates are site and tumor specific.
}
