% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_rate_calc.R
\name{mutation_rate_calc}
\alias{mutation_rate_calc}
\title{Mutation rate calculation}
\usage{
mutation_rate_calc(MAF, gene, gene_mut_rate, trinuc_proportion_matrix,
  gene_trinuc_comp, RefCDS)
}
\arguments{
\item{MAF}{The MAF file to extract mutational data from}

\item{gene}{Gene name in question}

\item{gene_mut_rate}{}

\item{RefCDS}{RefCDS loaded in from data("RefCDS_TP53splice",package = "cancereffectsizeR")}
}
\description{
This function
}
