% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_rate_calc.R
\name{mutation_rate_calc}
\alias{mutation_rate_calc}
\title{Mutation rate calculation}
\usage{
mutation_rate_calc(
  this_MAF,
  gene,
  gene_mut_rate,
  trinuc_proportion_matrix,
  all_tumors,
  progressions
)
}
\arguments{
\item{this_MAF}{The subset MAF (just the gene to be analyzed) file to extract mutational data from}

\item{gene}{current gene}

\item{gene_mut_rate}{mutation rate at the gene-level}

\item{trinuc_proportion_matrix}{matrix constructed from deconstructSigs output, containing proportion of each trinucleotide mutated in each tumor}

\item{progression}{CESProgressions}
}
\value{

}
\description{
This function calculates the mutation rates of all the unique varaints within a gene given gene-level mutation rate and tumor-specific trinucleotide contexts.
}
