% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_rate_calc.R
\name{mutation_rate_calc}
\alias{mutation_rate_calc}
\title{Mutation rate calculation}
\usage{
mutation_rate_calc(MAF, gene, gene_mut_rate, trinuc_proportion_matrix,
  gene_trinuc_comp, RefCDS,
  relative_substitution_rate = relative_substitution_rate)
}
\arguments{
\item{MAF}{The MAF file to extract mutational data from}

\item{gene}{Gene name in question}

\item{gene_mut_rate}{mutation rate at the gene-level}

\item{trinuc_proportion_matrix}{matrix constructed from deconstructSigs output, containing proportion of each trinucleotide mutated in each tumor}

\item{gene_trinuc_comp}{list containing matrices of counts of each trinuc in every gene}

\item{RefCDS}{RefCDS loaded in from data("RefCDS_TP53splice",package = "cancereffectsizeR")}
}
\description{
This function calculates the mutation rates of all the unique varaints within a gene given gene-level mutation rate and tumor-specific trinucleotide contexts.
}
