% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutational_signature_effects.R
\name{mutational_signature_effects}
\alias{mutational_signature_effects}
\title{Attribute cancer effects to mutational signatures}
\usage{
mutational_signature_effects(cesa = cesa, effects = NULL, samples = NULL)
}
\arguments{
\item{cesa}{CESAnalysis with cancer effects calculated for the variants of interest.}

\item{effects}{A table of cancer effect estimates for a set of variants, as produced with
\code{ces_variant()}. Different sets of variants (or parameter choices in the
\code{ces_variant} run) will affect output.}

\item{samples}{Samples for which to calculate mutational sources and effect shares; defaults to all
samples. Only included samples will inform the cohort-level output.}
}
\value{
A nested list containing mutational source probabilities for each variant (per patient
  and averaged over patients with the variant) and cancer effect shares attributable to each
  signature (per patient and across the cohort).
}
\description{
Within patients and across the cohort, calculates mutational source probabilities and the share
  of cancer effects attributable to each source, where sources are biologically-associated
  mutational signatures. See \href{https://doi.org/10.1093/molbev/msac084}{Attribution of Cancer Origins to Endogenous, Exogenous, and Preventable Mutational Processes}
  for background and applications.
}
