% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_epistasis.R
\name{pairwise_gene_epistasis}
\alias{pairwise_gene_epistasis}
\title{Calculate SIs at gene level under pairwise epistasis model}
\usage{
pairwise_gene_epistasis(cesa, genes)
}
\description{
The genes are assumed to not overlap in any ranges (the calling
function checks for this).
}
\keyword{internal}
