% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_variant_epistasis.R
\name{pairwise_variant_epistasis}
\alias{pairwise_variant_epistasis}
\title{Estimate selection under pairwise epistasis model}
\usage{
pairwise_variant_epistasis(
  cesa,
  variant_pair,
  samples,
  conf,
  compound_variants = NULL,
  model = "default",
  lik_args = list(),
  optimizer_args = list(),
  pval_calc_fn = NULL
)
}
\arguments{
\item{cesa}{CESAnalysis}

\item{variant_pair}{2-length character of variant IDs, or 2-length numeric giving
indices of VariantSetList for the current two compound variants}

\item{samples}{Validated samples data.table (as from select_samples())}

\item{compound_variants}{If testing a pair of compound variants, the VariantSetList defining them}

\item{model}{Passed from ces_epistasis or ces_gene_epistasis. Set to "default" to use built-in
model of epistatic selection, or supply a custom function factory (see details).}

\item{lik_args}{Extra arguments, given as a list, passed from ces_epistasis or ces_gene_epistasis}

\item{optimizer_args}{List of arguments to pass to the optimizer (bbmle::mle2).}

\item{pval_calc_fn}{For use with custom models; optional. A function that takes an epistasis
model fit as input and returns p-values and other descriptives.}
}
\description{
Estimate selection under pairwise epistasis model
}
\keyword{internal}
