% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_deconstructSigs.R
\name{run_deconstructSigs}
\alias{run_deconstructSigs}
\title{deconstructSigs wrapper}
\usage{
run_deconstructSigs(
  tumor_trinuc_counts,
  signatures_df,
  signatures_to_remove,
  artifact_signatures
)
}
\arguments{
\item{tumor_trinuc_counts}{one-row data.frame of trinuc variants counts (in deconstructSigs order) for one tumor}

\item{signatures_df}{data.frame of signatures (see COSMIC v3 signatures included with package for format)}

\item{signatures_to_remove}{names of signatures in signatures_df to keep out of deconstructSigs and assign zero weights}

\item{artifact_signatures}{names of signatures that should be treated as sequencing artifacts  (these have weights calculated
but are then normalized out when calculating true trinuc proportions)}
}
\description{
gets called internally by trinucleotide_mutation_weights function (don't call directly without a good reason)
}
