% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_handling.R
\name{select_variants}
\alias{select_variants}
\title{Select and filter variants}
\usage{
select_variants(
  cesa,
  genes = NULL,
  variant_ids = NULL,
  granges = NULL,
  min_freq = 1,
  include_subvariants = F,
  gr_padding = 0
)
}
\arguments{
\item{cesa}{CESAnalysis with MAF data loaded and annotated (e.g., with \code{load_maf()})}

\item{genes}{include variants (coding and noncoding) within these genes}

\item{variant_ids}{include variants by ID (vector of snv_id and/or aac_id)}

\item{granges}{filter out any variants not within input GRanges}

\item{min_freq}{filter out variants that occur in fewer than this many samples in the
MAF data Defaults to 1; set to 0 to include all passing variants that are in the
annotation tables. Note that variants that are not in the annotation tables will
never be returned. Use \code{generate_variants()} to annotate variants absent from the
data.}

\item{include_subvariants}{Some mutations "contain" other mutations. For example, in
cancereffectsizeR's default hg19 reference data, KRAS_Q61H contains two constituent
SNVs that both cause the same amino acid change: 12:25380275_T>G and 12:25380275_T>A.
When include_subvariants = F (the default), and genes = "KRAS", output will be
returned for KRAS_Q61H but not for the two SNVs (although their IDs will appear in
the Q61H output). Set to true, and all three variants will be included in output, 
assuming they don't get filtered out.}

\item{gr_padding}{add +/- this many bases to every range specified in "granges"
(stopping at chromsome ends, naturally)}
}
\description{
This is a lightweight function to help you find and manipulate variant data from your
CESAnalysis's MAF data and mutation annotation tables. This function has a variant
gathering step followed by a filtering step. For gathering, you can specify genes by
name or supply CES-style variant IDs to get all variants matching either. Or, leave
both empty and all available variants will be gathered. For filtering, use the granges
parameter to filter out variants that don't intersect an input GRanges object. Use
min_freq to filter out variants by frequency in the MAF data.
}
\details{
To collect all available variant data, set min_freq = 0, include_subvariants = T, and
no other options. Output will include additional rows for SNVs that are already
annotated as constituent SNVs of amino-acid-change mutations. Note that while
intergenic SNVs have their nearest genes annotated in the SNV tables, these variants
will not be captured by gene-based selection with this function, since they're not
actually in any gene.
}
