% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_handling.R
\name{select_variants}
\alias{select_variants}
\title{Select and filter variants}
\usage{
select_variants(
  cesa,
  genes = NULL,
  variant_ids = NULL,
  granges = NULL,
  min_freq = 1,
  include_subvariants = F
)
}
\arguments{
\item{cesa}{CESAnalysis with MAF data loaded and annotated (e.g., with \code{load_maf()})}

\item{genes}{include variants (coding and noncoding) within these genes}

\item{variant_ids}{include variants}

\item{granges}{include all variants overlapping this GRanges object}

\item{min_freq}{filter out variants that occur in fewer than this many samples in the
MAF data (Defaults to 1; set to 0 to include all passing variants that are in the
annotation tables. Note that variants that are not in the annotation tables will
never be returned. Use \code{generate_variants()} to annotate variants absent from the
data.}

\item{include_subvariants}{Some mutations "contain" other mutations. For example, in
cancereffectsizeR's, default hg19 reference data, KRAS_Q61H contains two constituent
SNVs that both cause the same amino acid change: 12:25380275_T>G and 12:25380275_T>A.
When include_subvariants = F (the default), and genes = "KRAS", output will be
returned for KRAS_Q61H but not for the two SNVs (although their IDs will appear in
the Q61H output). Set to true, and all three variants will be included in output, 
assuming they don't get filtered out by min_freq or otherwise.}

\item{gr_padding}{add +/- this many bases to every range specified in "granges"
(stopping at chromsome ends, naturally)}
}
\description{
This is a lightweight function to help you search for and organization variant data
that is present in your CESAnalysis's MAF data and mutation annotation tables. You can
specify genes by name as well as CES-style variant IDs, and all variants matching
either will be returned. Or, leave both empty and all available variants will be
returned. The frequency of each variant in the MAF data will also be reported, and you
can filter output on this frequency.
}
\details{
Note that while intergenic SNVs have their nearest genes annotated in the SNV tables,
these variants will not be captured by gene-based selection with this function, since
they're not actually in any gene.
}
