% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_variants.R
\name{select_variants}
\alias{select_variants}
\title{Select and filter variants}
\usage{
select_variants(
  cesa,
  genes = NULL,
  min_freq = 0,
  variant_ids = NULL,
  type = NULL,
  gr = NULL,
  variant_position_table = NULL,
  padding = 0
)
}
\arguments{
\item{cesa}{CESAnalysis with MAF data loaded and annotated (e.g., with \code{load_maf()})}

\item{genes}{Filter variants to specified genes.}

\item{min_freq}{Filter out variants with MAF frequency below threshold (default 0).
Note that variants that are not in the annotation tables will never be returned. Use
\code{add_variants()} to include variants absent from MAF data in your CESAnalysis.}

\item{variant_ids}{Filter out any variants besides those specified. (Variants specified here can
still get removed by other filters.)}

\item{type}{Filter out variants not of the specified types. Supported types are sbs, aac, dbs, dbs_aac.}

\item{gr}{Filter out any variants not within input GRanges +/- \code{padding} bases.}

\item{variant_position_table}{Filter out any variants that don't intersect the
positions given in chr/start/end of this table (1-based closed coordinates).
Typically, the table comes from a previous \code{select_variants} call and can be expanded
with \code{padding}. (Gritty detail: Amino acid substitutions get special handling. Only the
precise positions in start, end, and center_nt_pos are used. Otherwise, coding changes that
span splice sites would capture all the intronic sites.)}

\item{padding}{add +/- this many bases to every range specified in \code{gr} or
\code{variant_position_table} (stopping at chromosome ends, naturally).}
}
\value{
A data.table with info on selected variants (see details)
}
\description{
This function helps you find and view variant data from your CESAnalysis's MAF data and
mutation annotation tables. By default, substitutions that occur within coding regions are represented at
the codon level as amino acid changes (including synonymous substitutions, so amino acid "change"
is a bit of a misnomer), while substitutions outside of coding regions are represented as SBS. You can apply a series of filters to restrict output to
certain genes or genomic regions, require a minimum variant frequency in MAF data, and/or
specify exact \code{variant_ids} to return.
}
\details{
Only variants that are annotated in the CESAnalysis can be returned. To view more annotations,
such as for variants absent from the MAF data, you must first call add_variants() to add them to
the CESAnalysis.

Note that while intergenic SBS have their nearest genes annotated in the SBS
tables, these variants will not be captured by the genes filter of this function.

Definitions of some less self-explanatory columns:
\itemize{
  \item variant_name: In a coding variant, gene and protein change on the (MANE) canonical
  transcripts, such as "BRAF V600E". For coding changes reported on other transcripts, the
  protein ID is included: "POLH W415C (ENSP00000361300.1)". With older reference data sets
  (ces.refset.hg19, versions of ces.refset.hg38 < 1.3, and any custom reference data set that
  doesn't have complete information on canonical transcripts), the variant name is a 
  shortening of the variant_id.
  \item start/end: lowest/highest genomic positions overlapping variant
  \item variant_id: unique IDs for variants given the associated genome assembly version and the transcript data
  \item ref/alt: genomic reference and alternate alleles (for genomic
  variants; NA for AACs) 
  \item gene: the affected gene in AACs; for SBS, the
  overlapping gene (or an arbitrary gene if more than one overlaps), or the nearest gene
  for intergenic SBS
  \item strand: for AACs, 1 if the reference sequence strand is the coding strand; -1 otherwise
  \item essential_splice: Variant is 1,2 bp upstream or
  1,2,5 bp downstream of an annotated splice position (edge case: if an SBS has
  multiple gene/transcript annotations, this doesn't say which one it's essential for)
  \item intergenic: variant does not overlap any coding regions in the reference data
  \item trinuc_mut: for SBS, the reference trinucleotide context, in deconstructSigs notation
  \item coding_seq: coding strand nucleotides in order of transcription
  \item center_nt_pos: regardless of strand, start/end give positions of two out of three AAC nucleotides; this
                      gives the position of the center nucleotide (maybe useful if the AAC spans a splice site)
  \item maf_prevalence: number of occurrences of the variant in MAF data
  \item samples_covering: number of MAF samples with sequencing coverage at the variant site
}
}
