% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_results_converter.R
\name{selection_results_converter}
\alias{selection_results_converter}
\title{Selection results converter}
\usage{
selection_results_converter(cesa, min_recurrence = 2)
}
\arguments{
\item{cesa}{CESAnalysis object with selection intensity results}

\item{min_recurrence}{minimum number of samples in which a mutation must appear for the mutation to be included in results}
}
\value{

}
\description{
Selection results converter
}
\examples{

}
