% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CESAnalysis.R
\name{snv_results}
\alias{snv_results}
\title{View results from ces_variant}
\usage{
snv_results(cesa = NULL)
}
\arguments{
\item{cesa}{CESAnalysis object}
}
\description{
returns a list of ces_variant() results tables, with variant annotations added
}
