% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_cosmic_signatures_to_remove.R
\name{suggest_cosmic_signatures_to_remove}
\alias{suggest_cosmic_signatures_to_remove}
\title{Previous name of suggest_cosmic_signature_exclusions}
\usage{
suggest_cosmic_signatures_to_remove(...)
}
\description{
This function was renamed suggest_cosmic_signature_exclusions(). For now, calling it
with the old name will simply pass the arguments along to the new version. It will
eventually be removed.
}
\keyword{internal}
