% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trinuc_mutation_rates.R
\name{trinuc_mutation_rates}
\alias{trinuc_mutation_rates}
\title{Calculate relative rates of trinucleotide-context-specific mutations by extracting underlying mutational processes}
\usage{
trinuc_mutation_rates(
  cesa,
  signature_set = NULL,
  signatures_to_remove = character(),
  cores = 1,
  assume_identical_mutational_processes = FALSE,
  sig_averaging_threshold = 50,
  cosmic_hypermutation_rules = TRUE,
  artifact_accounting = TRUE,
  use_dS_exome2genome = FALSE
)
}
\arguments{
\item{cesa}{CESAnalysis object}

\item{signature_set}{name of built-in signature set (see \code{list_ces_signature_sets()}), or a custom signature set (see details)}

\item{signatures_to_remove}{specify any signatures to exclude from analysis; use \code{suggest_cosmic_signatures_to_remove()} for advice on COSMIC signatures}

\item{cores}{how many cores to use for processing tumors in parallel}

\item{assume_identical_mutational_processes}{use well-mutated tumors (those with number of eligible mutations meeting sig_averaging_threshold) 
to calculate group average signature weights, and assign these to all tumors}

\item{sig_averaging_threshold}{Mutational threshold (default 50) that determines which tumors inform the
calculation of group-average signature weightings. When assume_identical_mutational_processes == FALSE (the default), 
these group averages are blended into the signature weightings of tumors with few mutations (those below the threshold).}

\item{cosmic_hypermutation_rules}{T/F on whether to follow mutation count rules outlined in https://doi.org/10.1101/322859 (COSMIC v3 manuscript)
(only applies when running with COSMIC v3/v3.1 signatures)}

\item{artifact_accounting}{set false to disable special handling of artifact signatures (rarely recommended)}

\item{use_dS_exome2genome}{internal dev option (don't use)}
}
\value{
CESAnalysis with expected relative trinucleotide mutation rates ($trinuc_rates) and a table of tumor-specific 
        signature weights ($mutational_signatures). Note that tumors with group_avg_blended == TRUE have signature
        weights influenced by the average weights of well-mutated tumors; you may want to exclude these from some analyses.
}
\description{
This function calculates expected relative rates of trinucleotide-context-specific SNV
mutations within tumors by attributing SNVs to mutational processes represented
in mutation signature sets (such as COSMIC v3). deconstructSigs is used to extract mutational 
signature weights in each tumor. Tumors with targeted sequencing data are assigned the
average trinucleotide mutation rates calculated across all exome/genome data, 
which means that you need at least some exome or genome data to run.
}
\details{
To reduce the influence of selection on the estimation of relative trinucleotide mutation
rates, only non-recurrent SNVs (those that do not appear in more than one 
sample in the data set) are used.

A custom signature set should be given as a named three-item list, where "signatures"
is a pure data.frame with signature definitions, "name" is a 1-length character naming
the set, and "metadata" is a data.table with a "Signature" column that matches rownames
in the signature definitions. For artifact accounting to work, there should be a
TRUE/FALSE (logical) Likely_Artifact column; additional columns are optional. If you
don't have artifact information, you can also just supply an empty data.table. For a
template signature set object, run \code{sig_set = get_ces_signature_set("ces_hg19_v1",
"COSMIC_v3.1")}.
}
