% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trinucleotide_mutation_weights.R
\name{trinucleotide_mutation_weights}
\alias{trinucleotide_mutation_weights}
\title{Trinucleotide mutation weights}
\usage{
trinucleotide_mutation_weights(
  cesa,
  cores = 1,
  algorithm_choice = "weighted",
  remove_recurrent = TRUE,
  signature_choice = "cosmic_v3",
  v3_artifact_accounting = TRUE,
  v3_exome_hypermutation_rules = TRUE,
  signatures_to_remove = ""
)
}
\arguments{
\item{cesa}{# CESAnalysis object}

\item{cores}{how many cores to use to process tumor samples in parallel (requires parallel package)}

\item{algorithm_choice}{The choice of the algorithm used in determining trinucleotide weights.
Defaults to `weighted`, where all tumors with >= 50 substitutions have mutation
rates directly from deconstructSigs, and tumors with < 50 substitutions have that rate weighted
by the average of all tumors >= 50 substitutions relative to the substitution count < 50 in those tumors.
Other options currently include `all_average`, where all tumors have a mutation rate equal to the
average of all tumors >= 50 substitutions, `nearest_neighbor` where all tumors <50 substitutions
have the mutation rate of the tumor with >=50 substitutions closest to the mutational
distribution profile of the low substitution tumor as determined by a distance matrix, and
`all_calculated`, where all tumors have the mutation rate from the deconstructSigs output regardless of substitution number.}

\item{remove_recurrent}{if TRUE (default), removes recurrent variants from signature analysis}

\item{signature_choice}{"cosmic_v3" (default), "cosmic_v2", or a properly-formatted data frame with of trinucleotide signatures 
(if using cosmic_v3, just leave option default instead of passing your own data frame, or you'll get improper behavior)}

\item{v3_artifact_accounting}{when COSMIC v3 signatures associated with sequencing artifacts are detected, renormalizes to isolate "true" sources of mutational flux.}

\item{v3_exome_hypermutation_rules}{T/F on whether to follow the mutation count rules outlined in https://doi.org/10.1101/322859, the manuscript reported the v3 COSMIC signature set.}

\item{signatures_to_remove}{specify any signatures to exclude from analysis; some signatures automatically get excluded
from COSMIC v3 analyses; set to signatures_to_remove="none" to prevent this behavior (and )}
}
\description{
Calculates expected relative rates of SNV substitutions by trinucleotide context in tumors
using deconstructSigs to estimate weightings of known mutational signatures (e.g., COSMIC v3)
}
\details{
The purpose of this function is to calculate relative rates of trinucleotide-context-specific SNV
mutations within tumors that can be attributed to known tumor-specific mutational processes. Since
recurrent mutations (those that appear in multiple tumor samples) are more likely to be undergoing
selection than other mutations, by default only non-recurrent mutations are considered in order to
get a closer estimation of mutation rates independent of selection. This function currently uses
the deconstructSigs to assign mutational signature weightings to tumors.
As deconstructSigs suggests that a tumor must have at least 50 mutations for signature attribution
to be reliable, this function offers several methods for dealing with tumors with fewer mutations.
The default method, "weighted".... (to be continued)
}
