% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trinucleotide_mutation_weights.R
\name{trinucleotide_mutation_weights}
\alias{trinucleotide_mutation_weights}
\title{Trinucleotide mutation weights}
\usage{
trinucleotide_mutation_weights(cesa, algorithm_choice = "weighted",
  remove_recurrent = TRUE,
  signature_choice = "signatures_cosmic_May2019")
}
\arguments{
\item{cesa}{# CESAnalysis object}

\item{algorithm_choice}{The choice of the algorithm used in determining trinucleotide weights.
Defaults to `weighted`, where all tumors with >= 50 substitutions have mutation
rates directly from deconstructSigs, and tumors with < 50 substitutions have that rate weighted
by the average of all tumors >= 50 substitutions relative to the substitution count < 50 in those tumors.
Other options currently include `all_average`, where all tumors have a mutation rate equal to the
average of all tumors >= 50 substitutions, `nearest_neighbor` where all tumors <50 substitutions
have the mutation rate of the tumor with >=50 substitutions closest to the mutational
distribution profile of the low substitution tumor as determined by a distance matrix, and
`all_calculated`, where all tumors have the mutation rate from the deconstructSigs output regardless of substitution number.}

\item{remove_recurrent}{If TRUE, removes recurrent variants prior to calculating trinucleotide signatures and weights. If FALSE, uses all variants.}

\item{signature_choice}{Either "signatures_cosmic_May2019" (default) or "signatures.cosmic" (COSMIC signatures v2 originally packaged with deconstructSigs).}
}
\description{
Trinucleotide mutation weights
}
\examples{



only want to find the signatures that reflect the underlying mutational
processes within the tumor, i.e., those that reach detectable levels neutrally
so we detect the signatues from non-recurrently substituted variants
TODO: it is possible that the only mutation in a tumor is a recurrent variant,
and if we remove this variant, then the tumor is lost from the mutation
rate calculation. Should this be added back in if algorithm_choice
is something involving averages? ---- UPDATE: it is added in "weightd", need to
add in other options

TODO: really, averages should be averages of weight, and then product is calculated
individually per tumor, as opposed to average product and average rate.
}
