% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumor_allele_adder.R
\name{tumor_allele_adder}
\alias{tumor_allele_adder}
\title{Tumor allele column addition}
\usage{
tumor_allele_adder(MAF)
}
\arguments{
\item{MAF}{MAF file with columns labels "Reference_Allele",
"Tumor_Seq_Allele1", and "Tumor_Seq_Allele2"}
}
\description{
Several MAF files recently encountered had different representations of
the "tumor allele" . This function attempts to normalize all of the
different representations and put the tumor allele into a column called
"Tumor_allele".
}
