% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_handling.R
\name{validate_snv_ids}
\alias{validate_snv_ids}
\title{validate_snv_ids}
\usage{
validate_snv_ids(snv_ids, bsg)
}
\arguments{
\item{snv_ids}{character vector of snv_ids}

\item{bsg}{BSgenome for getting reference sequence}
}
\description{
Ensures SNV IDs are valid for the given genome
}
\keyword{internal}
